Split SSL hybrid-handshake + encoding test
