+++ ssl/CMakeLists.txt
@@ -88,7 +88,9 @@ if(BUILD_TESTING)
     ssl_misc_test.cc
     ssl_version_test.cc
     ssl_ech_test.cc
-          ssl_quic_test.cc
+    ssl_quic_test.cc
+    ssl_hybrid_handshake_test.cc
+    ssl_encoding_test.cc
   )
 
   target_link_libraries(${SSL_TEST_EXEC} boringssl_gtest_main ssl)
