+++ ssl/ssl_test.cc
@@ -99,20 +99,7 @@ struct CurveTest {
 
 
 
-struct HybridHandshakeTest {
-  // The curves rule string to apply to the client
-  const char *client_rule;
-  // TLS version that the client is configured with
-  uint16_t client_version;
-  // The curves rule string to apply to the server
-  const char *server_rule;
-  // TLS version that the server is configured with
-  uint16_t server_version;
-  // The group that is expected to be negotiated
-  uint16_t expected_group;
-  // Is a HelloRetryRequest expected?
-  bool is_hrr_expected;
-};
+
 
 
 
@@ -606,662 +593,6 @@ static const char *kBadCurvesLists[] = {
   "x25519_mlkem768:prime256v1",
 };
 
-static const HybridHandshakeTest kHybridHandshakeTests[] = {
-  // The corresponding hybrid group should be negotiated when client
-  // and server support only that group
-  {
-    "X25519Kyber768Draft00",
-    TLS1_3_VERSION,
-    "X25519Kyber768Draft00",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519_KYBER768_DRAFT00,
-    false,
-  },
-
-  {
-    "SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    "SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1_KYBER768_DRAFT00,
-    false,
-  },
-
-  // The client's preferred hybrid group should be negotiated when also
-  // supported by the server, even if the server "prefers"/supports other groups.
-  {
-    "X25519Kyber768Draft00:x25519",
-    TLS1_3_VERSION,
-    "x25519:prime256v1:X25519Kyber768Draft00",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519_KYBER768_DRAFT00,
-    false,
-  },
-
-  {
-    "X25519Kyber768Draft00:x25519",
-    TLS1_3_VERSION,
-    "X25519Kyber768Draft00:x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519_KYBER768_DRAFT00,
-    false,
-  },
-
-  {
-    "SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    "X25519Kyber768Draft00:secp384r1:x25519:SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1_KYBER768_DRAFT00,
-    false,
-  },
-
-  // The client lists PQ/hybrid groups as both first and second preferences.
-  // The key share logic is implemented such that the client will always
-  // attempt to send one hybrid key share and one classical key share.
-  // Therefore, the client will send key shares [SecP256r1Kyber768Draft00, x25519],
-  // skipping X25519Kyber768Draft00, and the server will choose to negotiate
-  // x25519 since it is the only mutually supported group.
-  {
-    "SecP256r1Kyber768Draft00:X25519Kyber768Draft00:x25519",
-    TLS1_3_VERSION,
-    "secp384r1:x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519,
-    false,
-  },
-
-  // The client will send key shares [x25519, SecP256r1Kyber768Draft00].
-  // The server will negotiate SecP256r1Kyber768Draft00 since it is the only
-  // mutually supported group.
-  {
-    "x25519:secp384r1:SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    "SecP256r1Kyber768Draft00:prime256v1",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1_KYBER768_DRAFT00,
-    false,
-  },
-
-  // The client will send key shares [x25519, SecP256r1Kyber768Draft00]. The
-  // server will negotiate x25519 since the client listed it as its first
-  // preference, even though it supports SecP256r1Kyber768Draft00.
-  {
-    "x25519:prime256v1:SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    "prime256v1:x25519:SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519,
-    false,
-  },
-
-  // The client will send key shares [SecP256r1Kyber768Draft00, x25519].
-  // The server will negotiate SecP256r1Kyber768Draft00 since the client listed
-  // it as its first preference.
-  {
-    "SecP256r1Kyber768Draft00:x25519:prime256v1",
-    TLS1_3_VERSION,
-    "prime256v1:x25519:SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1_KYBER768_DRAFT00,
-    false,
-  },
-
-  // In the supported_groups extension, the client will indicate its
-  // preferences, in order, as [SecP256r1Kyber768Draft00, X25519Kyber768Draft00,
-  // x25519, prime256v1]. From those groups, it will send key shares
-  // [SecP256r1Kyber768Draft00, x25519]. The server supports, and receives a
-  // key share for, x25519. However, when selecting a mutually supported group
-  // to negotiate, the server recognizes that the client prefers
-  // X25519Kyber768Draft00 over x25519. Since the server also supports
-  // X25519Kyber768Draft00, but did not receive a key share for it, it will
-  // select it and send an HRR. This ensures that the client's highest
-  // preference group will be negotiated, even at the expense of an additional
-  // round-trip.
-  //
-  // In our SSL implementation, this situation is unique to the case where the
-  // client supports both ECC and hybrid/PQ. When sending key shares, the
-  // client will send at most two key shares in one of the following ways:
-
-  // (a) one ECC key share - if the client supports only ECC;
-  // (b) one PQ key share - if the client supports only PQ;
-  // (c) one ECC and one PQ key share - if the client supports ECC and PQ.
-  //
-  // One of the above cases will be true irrespective of how many groups
-  // the client supports. If, say, the client supports four ECC groups
-  // and zero PQ groups, it will still only send a single ECC share. In cases
-  // (a) and (b), either the server supports that group and chooses to
-  // negotiate it, or it doesn't support it and sends an HRR. Case (c) is the
-  // only case where the server might receive a key share for a mutually
-  // supported group, but chooses to respect the client's preference order
-  // defined in the supported_groups extension at the expense of an additional
-  // round-trip.
-  {
-    "SecP256r1Kyber768Draft00:X25519Kyber768Draft00:x25519:prime256v1",
-    TLS1_3_VERSION,
-    "X25519Kyber768Draft00:prime256v1:x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519_KYBER768_DRAFT00,
-    true,
-  },
-
-  // Like the previous case, but the client's prioritization of ECC and PQ
-  // is inverted.
-  {
-    "x25519:prime256v1:SecP256r1Kyber768Draft00:X25519Kyber768Draft00",
-    TLS1_3_VERSION,
-    "X25519Kyber768Draft00:prime256v1",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1,
-    true,
-  },
-
-  // The client will send key shares [SecP256r1Kyber768Draft00, x25519]. The
-  // server will negotiate X25519Kyber768Draft00 after an HRR.
-  {
-    "SecP256r1Kyber768Draft00:X25519Kyber768Draft00:x25519:prime256v1",
-    TLS1_3_VERSION,
-    "X25519Kyber768Draft00:prime256v1",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519_KYBER768_DRAFT00,
-    true,
-  },
-
-  // EC should be negotiated when client prefers EC, or server does not
-  // support hybrid
-  {
-    "X25519Kyber768Draft00:x25519",
-    TLS1_3_VERSION,
-    "x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519,
-    false,
-  },
-  {
-    "x25519:SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    "x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519,
-    false,
-  },
-  {
-    "prime256v1:X25519Kyber768Draft00",
-    TLS1_3_VERSION,
-    "X25519Kyber768Draft00:prime256v1",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1,
-    false,
-  },
-  {
-    "prime256v1:x25519:SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    "x25519:prime256v1:SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1,
-    false,
-  },
-
-  // EC should be negotiated, after a HelloRetryRequest, if the server
-  // supports only curves for which it did not initially receive a key share
-  {
-    "X25519Kyber768Draft00:x25519:SecP256r1Kyber768Draft00:prime256v1",
-    TLS1_3_VERSION,
-    "prime256v1",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1,
-    true,
-  },
-  {
-    "X25519Kyber768Draft00:SecP256r1Kyber768Draft00:prime256v1:x25519",
-    TLS1_3_VERSION,
-    "secp224r1:secp384r1:secp521r1:x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519,
-    true,
-  },
-
-  // Hybrid should be negotiated, after a HelloRetryRequest, if the server
-  // supports only curves for which it did not initially receive a key share
-  {
-    "x25519:prime256v1:SecP256r1Kyber768Draft00:X25519Kyber768Draft00",
-    TLS1_3_VERSION,
-    "secp224r1:X25519Kyber768Draft00:secp521r1",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519_KYBER768_DRAFT00,
-    true,
-  },
-  {
-    "X25519Kyber768Draft00:x25519:prime256v1:SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    "SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1_KYBER768_DRAFT00,
-    true,
-  },
-
-  // If there is no overlap between client and server groups,
-  // the handshake should fail
-  {
-    "SecP256r1Kyber768Draft00:X25519Kyber768Draft00:secp384r1",
-    TLS1_3_VERSION,
-    "prime256v1:x25519",
-    TLS1_3_VERSION,
-    0,
-    false,
-  },
-  {
-    "secp384r1:SecP256r1Kyber768Draft00:X25519Kyber768Draft00",
-    TLS1_3_VERSION,
-    "prime256v1:x25519",
-    TLS1_3_VERSION,
-    0,
-    false,
-  },
-  {
-    "secp384r1:SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    "prime256v1:x25519:X25519Kyber768Draft00",
-    TLS1_3_VERSION,
-    0,
-    false,
-  },
-  {
-    "SecP256r1Kyber768Draft00",
-    TLS1_3_VERSION,
-    "X25519Kyber768Draft00",
-    TLS1_3_VERSION,
-    0,
-    false,
-  },
-
-  // If the client supports hybrid TLS 1.3, but the server
-  // only supports TLS 1.2, then TLS 1.2 EC should be negotiated.
-  {
-    "SecP256r1Kyber768Draft00:prime256v1",
-    TLS1_3_VERSION,
-    "prime256v1:x25519",
-    TLS1_2_VERSION,
-    SSL_GROUP_SECP256R1,
-    false,
-  },
-
-  // Same as above, but server also has SecP256r1Kyber768Draft00 in it's
-  // supported list, but can't use it since TLS 1.3 is the minimum version that
-  // supports PQ.
-  {
-    "SecP256r1Kyber768Draft00:prime256v1",
-    TLS1_3_VERSION,
-    "SecP256r1Kyber768Draft00:prime256v1:x25519",
-    TLS1_2_VERSION,
-    SSL_GROUP_SECP256R1,
-    false,
-  },
-
-  // If the client configures the curve list to include a hybrid
-  // curve, then initiates a 1.2 handshake, it will not advertise
-  // hybrid groups because hybrid is not supported for 1.2. So
-  // a 1.2 EC handshake will be negotiated (even if the server
-  // supports 1.3 with corresponding hybrid group).
-  {
-    "SecP256r1Kyber768Draft00:x25519",
-    TLS1_2_VERSION,
-    "SecP256r1Kyber768Draft00:x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519,
-    false,
-  },
-  {
-    "SecP256r1Kyber768Draft00:prime256v1",
-    TLS1_2_VERSION,
-    "prime256v1:x25519",
-    TLS1_2_VERSION,
-    SSL_GROUP_SECP256R1,
-    false,
-  },
-  // The corresponding hybrid group should be negotiated when client
-  // and server support only that group
-  {
-    "X25519MLKEM768",
-    TLS1_3_VERSION,
-    "X25519MLKEM768",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519_MLKEM768,
-    false,
-  },
-
-  {
-    "SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    "SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1_MLKEM768,
-    false,
-  },
-
-  {
-    "SecP384r1MLKEM1024",
-    TLS1_3_VERSION,
-    "SecP384r1MLKEM1024",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP384R1_MLKEM1024,
-    false,
-  },
-
-  // The client's preferred hybrid group should be negotiated when also
-  // supported by the server, even if the server "prefers"/supports other groups.
-  {
-    "X25519MLKEM768:x25519",
-    TLS1_3_VERSION,
-    "x25519:prime256v1:X25519MLKEM768",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519_MLKEM768,
-    false,
-  },
-
-  {
-    "X25519MLKEM768:x25519",
-    TLS1_3_VERSION,
-    "X25519MLKEM768:x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519_MLKEM768,
-    false,
-  },
-
-  {
-    "SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    "X25519MLKEM768:secp384r1:x25519:SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1_MLKEM768,
-    false,
-  },
-
-  {
-    "X25519MLKEM768:SecP256r1MLKEM768:SecP384r1MLKEM1024",
-    TLS1_3_VERSION,
-    "SecP384r1MLKEM1024:SecP256r1MLKEM768:X25519MLKEM768",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519_MLKEM768,
-    false,
-  },
-
-  {
-    "SecP384r1MLKEM1024:SecP256r1MLKEM768:X25519MLKEM768",
-    TLS1_3_VERSION,
-    "X25519MLKEM768:SecP256r1MLKEM768:SecP384r1MLKEM1024",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP384R1_MLKEM1024,
-    false,
-  },
-
-  // The client lists PQ/hybrid groups as both first and second preferences.
-  // The key share logic is implemented such that the client will always
-  // attempt to send one hybrid key share and one classical key share.
-  // Therefore, the client will send key shares [SecP256r1MLKEM768, x25519],
-  // skipping X25519MLKEM768, and the server will choose to negotiate
-  // x25519 since it is the only mutually supported group.
-  {
-    "SecP256r1MLKEM768:X25519MLKEM768:x25519",
-    TLS1_3_VERSION,
-    "secp384r1:x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519,
-    false,
-  },
-
-  // The client will send key shares [x25519, SecP256r1MLKEM768].
-  // The server will negotiate SecP256r1MLKEM768 since it is the only
-  // mutually supported group.
-  {
-    "x25519:secp384r1:SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    "SecP256r1MLKEM768:prime256v1",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1_MLKEM768,
-    false,
-  },
-
-  // The client will send key shares [x25519, SecP256r1MLKEM768]. The
-  // server will negotiate x25519 since the client listed it as its first
-  // preference, even though it supports SecP256r1MLKEM768.
-  {
-    "x25519:prime256v1:SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    "prime256v1:x25519:SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519,
-    false,
-  },
-
-  // The client will send key shares [SecP256r1MLKEM768, x25519].
-  // The server will negotiate SecP256r1MLKEM768 since the client listed
-  // it as its first preference.
-  {
-    "SecP256r1MLKEM768:x25519:prime256v1",
-    TLS1_3_VERSION,
-    "prime256v1:x25519:SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1_MLKEM768,
-    false,
-  },
-
-  // In the supported_groups extension, the client will indicate its
-  // preferences, in order, as [SecP256r1MLKEM768, X25519MLKEM768,
-  // x25519, prime256v1]. From those groups, it will send key shares
-  // [SecP256r1MLKEM768, x25519]. The server supports, and receives a
-  // key share for, x25519. However, when selecting a mutually supported group
-  // to negotiate, the server recognizes that the client prefers
-  // X25519MLKEM768 over x25519. Since the server also supports
-  // X25519MLKEM768, but did not receive a key share for it, it will
-  // select it and send an HRR. This ensures that the client's highest
-  // preference group will be negotiated, even at the expense of an additional
-  // round-trip.
-  //
-  // In our SSL implementation, this situation is unique to the case where the
-  // client supports both ECC and hybrid/PQ. When sending key shares, the
-  // client will send at most two key shares in one of the following ways:
-
-  // (a) one ECC key share - if the client supports only ECC;
-  // (b) one PQ key share - if the client supports only PQ;
-  // (c) one ECC and one PQ key share - if the client supports ECC and PQ.
-  //
-  // One of the above cases will be true irrespective of how many groups
-  // the client supports. If, say, the client supports four ECC groups
-  // and zero PQ groups, it will still only send a single ECC share. In cases
-  // (a) and (b), either the server supports that group and chooses to
-  // negotiate it, or it doesn't support it and sends an HRR. Case (c) is the
-  // only case where the server might receive a key share for a mutually
-  // supported group, but chooses to respect the client's preference order
-  // defined in the supported_groups extension at the expense of an additional
-  // round-trip.
-  {
-    "SecP256r1MLKEM768:X25519MLKEM768:x25519:prime256v1",
-    TLS1_3_VERSION,
-    "X25519MLKEM768:prime256v1:x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519_MLKEM768,
-    true,
-  },
-
-  // Like the previous case, but the client's prioritization of ECC and PQ
-  // is inverted.
-  {
-    "x25519:prime256v1:SecP256r1MLKEM768:X25519MLKEM768",
-    TLS1_3_VERSION,
-    "X25519MLKEM768:prime256v1",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1,
-    true,
-  },
-
-  // The client will send key shares [SecP256r1MLKEM768, x25519]. The
-  // server will negotiate X25519MLKEM768 after an HRR.
-  {
-    "SecP256r1MLKEM768:X25519MLKEM768:x25519:prime256v1",
-    TLS1_3_VERSION,
-    "X25519MLKEM768:prime256v1",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519_MLKEM768,
-    true,
-  },
-
-  // EC should be negotiated when client prefers EC, or server does not
-  // support hybrid
-  {
-    "X25519MLKEM768:x25519",
-    TLS1_3_VERSION,
-    "x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519,
-    false,
-  },
-  {
-    "x25519:SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    "x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519,
-    false,
-  },
-  {
-    "prime256v1:X25519MLKEM768",
-    TLS1_3_VERSION,
-    "X25519MLKEM768:prime256v1",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1,
-    false,
-  },
-  {
-    "prime256v1:x25519:SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    "x25519:prime256v1:SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1,
-    false,
-  },
-
-  // EC should be negotiated, after a HelloRetryRequest, if the server
-  // supports only curves for which it did not initially receive a key share
-  {
-    "X25519MLKEM768:x25519:SecP256r1MLKEM768:prime256v1",
-    TLS1_3_VERSION,
-    "prime256v1",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1,
-    true,
-  },
-  {
-    "X25519MLKEM768:SecP256r1MLKEM768:prime256v1:x25519",
-    TLS1_3_VERSION,
-    "secp224r1:secp384r1:secp521r1:x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519,
-    true,
-  },
-
-  // Hybrid should be negotiated, after a HelloRetryRequest, if the server
-  // supports only curves for which it did not initially receive a key share
-  {
-    "x25519:prime256v1:SecP256r1MLKEM768:X25519MLKEM768",
-    TLS1_3_VERSION,
-    "secp224r1:X25519MLKEM768:secp521r1",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519_MLKEM768,
-    true,
-  },
-  {
-    "X25519MLKEM768:x25519:prime256v1:SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    "SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    SSL_GROUP_SECP256R1_MLKEM768,
-    true,
-  },
-
-  // If there is no overlap between client and server groups,
-  // the handshake should fail
-  {
-    "SecP256r1MLKEM768:X25519MLKEM768:secp384r1",
-    TLS1_3_VERSION,
-    "prime256v1:x25519",
-    TLS1_3_VERSION,
-    0,
-    false,
-  },
-  {
-    "secp384r1:SecP256r1MLKEM768:X25519MLKEM768",
-    TLS1_3_VERSION,
-    "prime256v1:x25519",
-    TLS1_3_VERSION,
-    0,
-    false,
-  },
-  {
-    "secp384r1:SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    "prime256v1:x25519:X25519MLKEM768",
-    TLS1_3_VERSION,
-    0,
-    false,
-  },
-  {
-    "SecP256r1MLKEM768",
-    TLS1_3_VERSION,
-    "X25519MLKEM768",
-    TLS1_3_VERSION,
-    0,
-    false,
-  },
-
-  // If the client supports hybrid TLS 1.3, but the server
-  // only supports TLS 1.2, then TLS 1.2 EC should be negotiated.
-  {
-    "SecP256r1MLKEM768:prime256v1",
-    TLS1_3_VERSION,
-    "prime256v1:x25519",
-    TLS1_2_VERSION,
-    SSL_GROUP_SECP256R1,
-    false,
-  },
-
-  // Same as above, but server also has SecP256r1MLKEM768 in it's
-  // supported list, but can't use it since TLS 1.3 is the minimum version that
-  // supports PQ.
-  {
-    "SecP256r1MLKEM768:prime256v1",
-    TLS1_3_VERSION,
-    "SecP256r1MLKEM768:prime256v1:x25519",
-    TLS1_2_VERSION,
-    SSL_GROUP_SECP256R1,
-    false,
-  },
-
-  // If the client configures the curve list to include a hybrid
-  // curve, then initiates a 1.2 handshake, it will not advertise
-  // hybrid groups because hybrid is not supported for 1.2. So
-  // a 1.2 EC handshake will be negotiated (even if the server
-  // supports 1.3 with corresponding hybrid group).
-  {
-    "SecP256r1MLKEM768:x25519",
-    TLS1_2_VERSION,
-    "SecP256r1MLKEM768:x25519",
-    TLS1_3_VERSION,
-    SSL_GROUP_X25519,
-    false,
-  },
-  {
-    "SecP256r1MLKEM768:prime256v1",
-    TLS1_2_VERSION,
-    "prime256v1:x25519",
-    TLS1_2_VERSION,
-    SSL_GROUP_SECP256R1,
-    false,
-  },
-};
-
-
-
 static STACK_OF(SSL_CIPHER) *tls13_ciphers(const SSL_CTX *ctx) {
   return ctx->tls13_cipher_list->ciphers.get();
 }
@@ -1311,11 +642,6 @@ static bool CipherListsEqual(SSL_CTX *ctx,
 
   return true;
 }
-
-
-
-
-
 
 TEST(SSLTest, CipherRules) {
   for (const CipherTest &t : kCipherTests) {
@@ -1507,288 +833,7 @@ TEST(SSLTest, CurveRules) {
   }
 }
 
-// kOpenSSLSession is a serialized SSL_SESSION.
-static const char kOpenSSLSession[] =
-    "MIIFqgIBAQICAwMEAsAvBCAG5Q1ndq4Yfmbeo1zwLkNRKmCXGdNgWvGT3cskV0yQ"
-    "kAQwJlrlzkAWBOWiLj/jJ76D7l+UXoizP2KI2C7I2FccqMmIfFmmkUy32nIJ0mZH"
-    "IWoJoQYCBFRDO46iBAICASyjggR6MIIEdjCCA16gAwIBAgIIK9dUvsPWSlUwDQYJ"
-    "KoZIhvcNAQEFBQAwSTELMAkGA1UEBhMCVVMxEzARBgNVBAoTCkdvb2dsZSBJbmMx"
-    "JTAjBgNVBAMTHEdvb2dsZSBJbnRlcm5ldCBBdXRob3JpdHkgRzIwHhcNMTQxMDA4"
-    "MTIwNzU3WhcNMTUwMTA2MDAwMDAwWjBoMQswCQYDVQQGEwJVUzETMBEGA1UECAwK"
-    "Q2FsaWZvcm5pYTEWMBQGA1UEBwwNTW91bnRhaW4gVmlldzETMBEGA1UECgwKR29v"
-    "Z2xlIEluYzEXMBUGA1UEAwwOd3d3Lmdvb2dsZS5jb20wggEiMA0GCSqGSIb3DQEB"
-    "AQUAA4IBDwAwggEKAoIBAQCcKeLrplAC+Lofy8t/wDwtB6eu72CVp0cJ4V3lknN6"
-    "huH9ct6FFk70oRIh/VBNBBz900jYy+7111Jm1b8iqOTQ9aT5C7SEhNcQFJvqzH3e"
-    "MPkb6ZSWGm1yGF7MCQTGQXF20Sk/O16FSjAynU/b3oJmOctcycWYkY0ytS/k3LBu"
-    "Id45PJaoMqjB0WypqvNeJHC3q5JjCB4RP7Nfx5jjHSrCMhw8lUMW4EaDxjaR9KDh"
-    "PLgjsk+LDIySRSRDaCQGhEOWLJZVLzLo4N6/UlctCHEllpBUSvEOyFga52qroGjg"
-    "rf3WOQ925MFwzd6AK+Ich0gDRg8sQfdLH5OuP1cfLfU1AgMBAAGjggFBMIIBPTAd"
-    "BgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwGQYDVR0RBBIwEIIOd3d3Lmdv"
-    "b2dsZS5jb20waAYIKwYBBQUHAQEEXDBaMCsGCCsGAQUFBzAChh9odHRwOi8vcGtp"
-    "Lmdvb2dsZS5jb20vR0lBRzIuY3J0MCsGCCsGAQUFBzABhh9odHRwOi8vY2xpZW50"
-    "czEuZ29vZ2xlLmNvbS9vY3NwMB0GA1UdDgQWBBQ7a+CcxsZByOpc+xpYFcIbnUMZ"
-    "hTAMBgNVHRMBAf8EAjAAMB8GA1UdIwQYMBaAFErdBhYbvPZotXb1gba7Yhq6WoEv"
-    "MBcGA1UdIAQQMA4wDAYKKwYBBAHWeQIFATAwBgNVHR8EKTAnMCWgI6Ahhh9odHRw"
-    "Oi8vcGtpLmdvb2dsZS5jb20vR0lBRzIuY3JsMA0GCSqGSIb3DQEBBQUAA4IBAQCa"
-    "OXCBdoqUy5bxyq+Wrh1zsyyCFim1PH5VU2+yvDSWrgDY8ibRGJmfff3r4Lud5kal"
-    "dKs9k8YlKD3ITG7P0YT/Rk8hLgfEuLcq5cc0xqmE42xJ+Eo2uzq9rYorc5emMCxf"
-    "5L0TJOXZqHQpOEcuptZQ4OjdYMfSxk5UzueUhA3ogZKRcRkdB3WeWRp+nYRhx4St"
-    "o2rt2A0MKmY9165GHUqMK9YaaXHDXqBu7Sefr1uSoAP9gyIJKeihMivsGqJ1TD6Z"
-    "cc6LMe+dN2P8cZEQHtD1y296ul4Mivqk3jatUVL8/hCwgch9A8O4PGZq9WqBfEWm"
-    "IyHh1dPtbg1lOXdYCWtjpAIEAKUDAgEUqQUCAwGJwKqBpwSBpBwUQvoeOk0Kg36S"
-    "YTcLEkXqKwOBfF9vE4KX0NxeLwjcDTpsuh3qXEaZ992r1N38VDcyS6P7I6HBYN9B"
-    "sNHM362zZnY27GpTw+Kwd751CLoXFPoaMOe57dbBpXoro6Pd3BTbf/Tzr88K06yE"
-    "OTDKPNj3+inbMaVigtK4PLyPq+Topyzvx9USFgRvyuoxn0Hgb+R0A3j6SLRuyOdA"
-    "i4gv7Y5oliyntgMBAQA=";
 
-// kCustomSession is a custom serialized SSL_SESSION generated by
-// filling in missing fields from |kOpenSSLSession|. This includes
-// providing |peer_sha256|, so |peer| is not serialized.
-static const char kCustomSession[] =
-    "MIIBZAIBAQICAwMEAsAvBCAG5Q1ndq4Yfmbeo1zwLkNRKmCXGdNgWvGT3cskV0yQ"
-    "kAQwJlrlzkAWBOWiLj/jJ76D7l+UXoizP2KI2C7I2FccqMmIfFmmkUy32nIJ0mZH"
-    "IWoJoQYCBFRDO46iBAICASykAwQBAqUDAgEUqAcEBXdvcmxkqQUCAwGJwKqBpwSB"
-    "pBwUQvoeOk0Kg36SYTcLEkXqKwOBfF9vE4KX0NxeLwjcDTpsuh3qXEaZ992r1N38"
-    "VDcyS6P7I6HBYN9BsNHM362zZnY27GpTw+Kwd751CLoXFPoaMOe57dbBpXoro6Pd"
-    "3BTbf/Tzr88K06yEOTDKPNj3+inbMaVigtK4PLyPq+Topyzvx9USFgRvyuoxn0Hg"
-    "b+R0A3j6SLRuyOdAi4gv7Y5oliynrSIEIAYGBgYGBgYGBgYGBgYGBgYGBgYGBgYG"
-    "BgYGBgYGBgYGrgMEAQevAwQBBLADBAEF";
-
-// kBoringSSLSession is a serialized SSL_SESSION generated from bssl client.
-static const char kBoringSSLSession[] =
-    "MIIRwQIBAQICAwMEAsAvBCDdoGxGK26mR+8lM0uq6+k9xYuxPnwAjpcF9n0Yli9R"
-    "kQQwbyshfWhdi5XQ1++7n2L1qqrcVlmHBPpr6yknT/u4pUrpQB5FZ7vqvNn8MdHf"
-    "9rWgoQYCBFXgs7uiBAICHCCjggR6MIIEdjCCA16gAwIBAgIIf+yfD7Y6UicwDQYJ"
-    "KoZIhvcNAQELBQAwSTELMAkGA1UEBhMCVVMxEzARBgNVBAoTCkdvb2dsZSBJbmMx"
-    "JTAjBgNVBAMTHEdvb2dsZSBJbnRlcm5ldCBBdXRob3JpdHkgRzIwHhcNMTUwODEy"
-    "MTQ1MzE1WhcNMTUxMTEwMDAwMDAwWjBoMQswCQYDVQQGEwJVUzETMBEGA1UECAwK"
-    "Q2FsaWZvcm5pYTEWMBQGA1UEBwwNTW91bnRhaW4gVmlldzETMBEGA1UECgwKR29v"
-    "Z2xlIEluYzEXMBUGA1UEAwwOd3d3Lmdvb2dsZS5jb20wggEiMA0GCSqGSIb3DQEB"
-    "AQUAA4IBDwAwggEKAoIBAQC0MeG5YGQ0t+IeJeoneP/PrhEaieibeKYkbKVLNZpo"
-    "PLuBinvhkXZo3DC133NpCBpy6ZktBwamqyixAyuk/NU6OjgXqwwxfQ7di1AInLIU"
-    "792c7hFyNXSUCG7At8Ifi3YwBX9Ba6u/1d6rWTGZJrdCq3QU11RkKYyTq2KT5mce"
-    "Tv9iGKqSkSTlp8puy/9SZ/3DbU3U+BuqCFqeSlz7zjwFmk35acdCilpJlVDDN5C/"
-    "RCh8/UKc8PaL+cxlt531qoTENvYrflBno14YEZlCBZsPiFeUSILpKEj3Ccwhy0eL"
-    "EucWQ72YZU8mUzXBoXGn0zA0crFl5ci/2sTBBGZsylNBAgMBAAGjggFBMIIBPTAd"
-    "BgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwGQYDVR0RBBIwEIIOd3d3Lmdv"
-    "b2dsZS5jb20waAYIKwYBBQUHAQEEXDBaMCsGCCsGAQUFBzAChh9odHRwOi8vcGtp"
-    "Lmdvb2dsZS5jb20vR0lBRzIuY3J0MCsGCCsGAQUFBzABhh9odHRwOi8vY2xpZW50"
-    "czEuZ29vZ2xlLmNvbS9vY3NwMB0GA1UdDgQWBBS/bzHxcE73Q4j3slC4BLbMtLjG"
-    "GjAMBgNVHRMBAf8EAjAAMB8GA1UdIwQYMBaAFErdBhYbvPZotXb1gba7Yhq6WoEv"
-    "MBcGA1UdIAQQMA4wDAYKKwYBBAHWeQIFATAwBgNVHR8EKTAnMCWgI6Ahhh9odHRw"
-    "Oi8vcGtpLmdvb2dsZS5jb20vR0lBRzIuY3JsMA0GCSqGSIb3DQEBCwUAA4IBAQAb"
-    "qdWPZEHk0X7iKPCTHL6S3w6q1eR67goxZGFSM1lk1hjwyu7XcLJuvALVV9uY3ovE"
-    "kQZSHwT+pyOPWQhsSjO+1GyjvCvK/CAwiUmBX+bQRGaqHsRcio7xSbdVcajQ3bXd"
-    "X+s0WdbOpn6MStKAiBVloPlSxEI8pxY6x/BBCnTIk/+DMB17uZlOjG3vbAnkDkP+"
-    "n0OTucD9sHV7EVj9XUxi51nOfNBCN/s7lpUjDS/NJ4k3iwOtbCPswiot8vLO779a"
-    "f07vR03r349Iz/KTzk95rlFtX0IU+KYNxFNsanIXZ+C9FYGRXkwhHcvFb4qMUB1y"
-    "TTlM80jBMOwyjZXmjRAhpAIEAKUDAgEUqQUCAwGJwKqBpwSBpOgebbmn9NRUtMWH"
-    "+eJpqA5JLMFSMCChOsvKey3toBaCNGU7HfAEiiXNuuAdCBoK262BjQc2YYfqFzqH"
-    "zuppopXCvhohx7j/tnCNZIMgLYt/O9SXK2RYI5z8FhCCHvB4CbD5G0LGl5EFP27s"
-    "Jb6S3aTTYPkQe8yZSlxevg6NDwmTogLO9F7UUkaYmVcMQhzssEE2ZRYNwSOU6KjE"
-    "0Yj+8fAiBtbQriIEIN2L8ZlpaVrdN5KFNdvcmOxJu81P8q53X55xQyGTnGWwsgMC"
-    "ARezggvvMIIEdjCCA16gAwIBAgIIf+yfD7Y6UicwDQYJKoZIhvcNAQELBQAwSTEL"
-    "MAkGA1UEBhMCVVMxEzARBgNVBAoTCkdvb2dsZSBJbmMxJTAjBgNVBAMTHEdvb2ds"
-    "ZSBJbnRlcm5ldCBBdXRob3JpdHkgRzIwHhcNMTUwODEyMTQ1MzE1WhcNMTUxMTEw"
-    "MDAwMDAwWjBoMQswCQYDVQQGEwJVUzETMBEGA1UECAwKQ2FsaWZvcm5pYTEWMBQG"
-    "A1UEBwwNTW91bnRhaW4gVmlldzETMBEGA1UECgwKR29vZ2xlIEluYzEXMBUGA1UE"
-    "AwwOd3d3Lmdvb2dsZS5jb20wggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIB"
-    "AQC0MeG5YGQ0t+IeJeoneP/PrhEaieibeKYkbKVLNZpoPLuBinvhkXZo3DC133Np"
-    "CBpy6ZktBwamqyixAyuk/NU6OjgXqwwxfQ7di1AInLIU792c7hFyNXSUCG7At8If"
-    "i3YwBX9Ba6u/1d6rWTGZJrdCq3QU11RkKYyTq2KT5mceTv9iGKqSkSTlp8puy/9S"
-    "Z/3DbU3U+BuqCFqeSlz7zjwFmk35acdCilpJlVDDN5C/RCh8/UKc8PaL+cxlt531"
-    "qoTENvYrflBno14YEZlCBZsPiFeUSILpKEj3Ccwhy0eLEucWQ72YZU8mUzXBoXGn"
-    "0zA0crFl5ci/2sTBBGZsylNBAgMBAAGjggFBMIIBPTAdBgNVHSUEFjAUBggrBgEF"
-    "BQcDAQYIKwYBBQUHAwIwGQYDVR0RBBIwEIIOd3d3Lmdvb2dsZS5jb20waAYIKwYB"
-    "BQUHAQEEXDBaMCsGCCsGAQUFBzAChh9odHRwOi8vcGtpLmdvb2dsZS5jb20vR0lB"
-    "RzIuY3J0MCsGCCsGAQUFBzABhh9odHRwOi8vY2xpZW50czEuZ29vZ2xlLmNvbS9v"
-    "Y3NwMB0GA1UdDgQWBBS/bzHxcE73Q4j3slC4BLbMtLjGGjAMBgNVHRMBAf8EAjAA"
-    "MB8GA1UdIwQYMBaAFErdBhYbvPZotXb1gba7Yhq6WoEvMBcGA1UdIAQQMA4wDAYK"
-    "KwYBBAHWeQIFATAwBgNVHR8EKTAnMCWgI6Ahhh9odHRwOi8vcGtpLmdvb2dsZS5j"
-    "b20vR0lBRzIuY3JsMA0GCSqGSIb3DQEBCwUAA4IBAQAbqdWPZEHk0X7iKPCTHL6S"
-    "3w6q1eR67goxZGFSM1lk1hjwyu7XcLJuvALVV9uY3ovEkQZSHwT+pyOPWQhsSjO+"
-    "1GyjvCvK/CAwiUmBX+bQRGaqHsRcio7xSbdVcajQ3bXdX+s0WdbOpn6MStKAiBVl"
-    "oPlSxEI8pxY6x/BBCnTIk/+DMB17uZlOjG3vbAnkDkP+n0OTucD9sHV7EVj9XUxi"
-    "51nOfNBCN/s7lpUjDS/NJ4k3iwOtbCPswiot8vLO779af07vR03r349Iz/KTzk95"
-    "rlFtX0IU+KYNxFNsanIXZ+C9FYGRXkwhHcvFb4qMUB1yTTlM80jBMOwyjZXmjRAh"
-    "MIID8DCCAtigAwIBAgIDAjqDMA0GCSqGSIb3DQEBCwUAMEIxCzAJBgNVBAYTAlVT"
-    "MRYwFAYDVQQKEw1HZW9UcnVzdCBJbmMuMRswGQYDVQQDExJHZW9UcnVzdCBHbG9i"
-    "YWwgQ0EwHhcNMTMwNDA1MTUxNTU2WhcNMTYxMjMxMjM1OTU5WjBJMQswCQYDVQQG"
-    "EwJVUzETMBEGA1UEChMKR29vZ2xlIEluYzElMCMGA1UEAxMcR29vZ2xlIEludGVy"
-    "bmV0IEF1dGhvcml0eSBHMjCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEB"
-    "AJwqBHdc2FCROgajguDYUEi8iT/xGXAaiEZ+4I/F8YnOIe5a/mENtzJEiaB0C1NP"
-    "VaTOgmKV7utZX8bhBYASxF6UP7xbSDj0U/ck5vuR6RXEz/RTDfRK/J9U3n2+oGtv"
-    "h8DQUB8oMANA2ghzUWx//zo8pzcGjr1LEQTrfSTe5vn8MXH7lNVg8y5Kr0LSy+rE"
-    "ahqyzFPdFUuLH8gZYR/Nnag+YyuENWllhMgZxUYi+FOVvuOAShDGKuy6lyARxzmZ"
-    "EASg8GF6lSWMTlJ14rbtCMoU/M4iarNOz0YDl5cDfsCx3nuvRTPPuj5xt970JSXC"
-    "DTWJnZ37DhF5iR43xa+OcmkCAwEAAaOB5zCB5DAfBgNVHSMEGDAWgBTAephojYn7"
-    "qwVkDBF9qn1luMrMTjAdBgNVHQ4EFgQUSt0GFhu89mi1dvWBtrtiGrpagS8wDgYD"
-    "VR0PAQH/BAQDAgEGMC4GCCsGAQUFBwEBBCIwIDAeBggrBgEFBQcwAYYSaHR0cDov"
-    "L2cuc3ltY2QuY29tMBIGA1UdEwEB/wQIMAYBAf8CAQAwNQYDVR0fBC4wLDAqoCig"
-    "JoYkaHR0cDovL2cuc3ltY2IuY29tL2NybHMvZ3RnbG9iYWwuY3JsMBcGA1UdIAQQ"
-    "MA4wDAYKKwYBBAHWeQIFATANBgkqhkiG9w0BAQsFAAOCAQEAqvqpIM1qZ4PtXtR+"
-    "3h3Ef+AlBgDFJPupyC1tft6dgmUsgWM0Zj7pUsIItMsv91+ZOmqcUHqFBYx90SpI"
-    "hNMJbHzCzTWf84LuUt5oX+QAihcglvcpjZpNy6jehsgNb1aHA30DP9z6eX0hGfnI"
-    "Oi9RdozHQZJxjyXON/hKTAAj78Q1EK7gI4BzfE00LshukNYQHpmEcxpw8u1VDu4X"
-    "Bupn7jLrLN1nBz/2i8Jw3lsA5rsb0zYaImxssDVCbJAJPZPpZAkiDoUGn8JzIdPm"
-    "X4DkjYUiOnMDsWCOrmji9D6X52ASCWg23jrW4kOVWzeBkoEfu43XrVJkFleW2V40"
-    "fsg12DCCA30wggLmoAMCAQICAxK75jANBgkqhkiG9w0BAQUFADBOMQswCQYDVQQG"
-    "EwJVUzEQMA4GA1UEChMHRXF1aWZheDEtMCsGA1UECxMkRXF1aWZheCBTZWN1cmUg"
-    "Q2VydGlmaWNhdGUgQXV0aG9yaXR5MB4XDTAyMDUyMTA0MDAwMFoXDTE4MDgyMTA0"
-    "MDAwMFowQjELMAkGA1UEBhMCVVMxFjAUBgNVBAoTDUdlb1RydXN0IEluYy4xGzAZ"
-    "BgNVBAMTEkdlb1RydXN0IEdsb2JhbCBDQTCCASIwDQYJKoZIhvcNAQEBBQADggEP"
-    "ADCCAQoCggEBANrMGGMw/fQXIxpWflvfPGw45HG3eJHUvKHYTPioQ7YD6U0hBwiI"
-    "2lgvZjkpvQV4i5046AW3an5xpObEYKaw74DkiSgPniXW7YPzraaRx5jJQhg1FJ2t"
-    "mEaSLk/K8YdDwRaVVy1Q74ktgHpXrfLuX2vSAI25FPgUFTXZwEaje3LIkb/JVSvN"
-    "0Jc+nCZkzN/Ogxlxyk7m1NV7qRnNVd7I7NJeOFPlXE+MLf5QIzb8ZubLjqQ5GQC3"
-    "lQI5kQsO/jgu0R0FmvZNPm8PBx2vLB6PYDni+jZTEznUXiYr2z2oFL0y6xgDKFIE"
-    "ceWrMz3hOLsHNoRinHnqFjD0X8Ar6HFr5PkCAwEAAaOB8DCB7TAfBgNVHSMEGDAW"
-    "gBRI5mj5K9KylddH2CMgEE8zmJCf1DAdBgNVHQ4EFgQUwHqYaI2J+6sFZAwRfap9"
-    "ZbjKzE4wDwYDVR0TAQH/BAUwAwEB/zAOBgNVHQ8BAf8EBAMCAQYwOgYDVR0fBDMw"
-    "MTAvoC2gK4YpaHR0cDovL2NybC5nZW90cnVzdC5jb20vY3Jscy9zZWN1cmVjYS5j"
-    "cmwwTgYDVR0gBEcwRTBDBgRVHSAAMDswOQYIKwYBBQUHAgEWLWh0dHBzOi8vd3d3"
-    "Lmdlb3RydXN0LmNvbS9yZXNvdXJjZXMvcmVwb3NpdG9yeTANBgkqhkiG9w0BAQUF"
-    "AAOBgQB24RJuTksWEoYwBrKBCM/wCMfHcX5m7sLt1Dsf//DwyE7WQziwuTB9GNBV"
-    "g6JqyzYRnOhIZqNtf7gT1Ef+i1pcc/yu2RsyGTirlzQUqpbS66McFAhJtrvlke+D"
-    "NusdVm/K2rxzY5Dkf3s+Iss9B+1fOHSc4wNQTqGvmO5h8oQ/Eg==";
-
-// kBadSessionExtraField is a custom serialized SSL_SESSION generated by
-// replacing the final (optional) element of |kCustomSession| with tag
-// number 99.
-static const char kBadSessionExtraField[] =
-    "MIIBdgIBAQICAwMEAsAvBCAG5Q1ndq4Yfmbeo1zwLkNRKmCXGdNgWvGT3cskV0yQ"
-    "kAQwJlrlzkAWBOWiLj/jJ76D7l+UXoizP2KI2C7I2FccqMmIfFmmkUy32nIJ0mZH"
-    "IWoJoQYCBFRDO46iBAICASykAwQBAqUDAgEUphAEDnd3dy5nb29nbGUuY29tqAcE"
-    "BXdvcmxkqQUCAwGJwKqBpwSBpBwUQvoeOk0Kg36SYTcLEkXqKwOBfF9vE4KX0Nxe"
-    "LwjcDTpsuh3qXEaZ992r1N38VDcyS6P7I6HBYN9BsNHM362zZnY27GpTw+Kwd751"
-    "CLoXFPoaMOe57dbBpXoro6Pd3BTbf/Tzr88K06yEOTDKPNj3+inbMaVigtK4PLyP"
-    "q+Topyzvx9USFgRvyuoxn0Hgb+R0A3j6SLRuyOdAi4gv7Y5oliynrSIEIAYGBgYG"
-    "BgYGBgYGBgYGBgYGBgYGBgYGBgYGBgYGBgYGrgMEAQevAwQBBOMDBAEF";
-
-// kBadSessionVersion is a custom serialized SSL_SESSION generated by replacing
-// the version of |kCustomSession| with 2.
-static const char kBadSessionVersion[] =
-    "MIIBdgIBAgICAwMEAsAvBCAG5Q1ndq4Yfmbeo1zwLkNRKmCXGdNgWvGT3cskV0yQ"
-    "kAQwJlrlzkAWBOWiLj/jJ76D7l+UXoizP2KI2C7I2FccqMmIfFmmkUy32nIJ0mZH"
-    "IWoJoQYCBFRDO46iBAICASykAwQBAqUDAgEUphAEDnd3dy5nb29nbGUuY29tqAcE"
-    "BXdvcmxkqQUCAwGJwKqBpwSBpBwUQvoeOk0Kg36SYTcLEkXqKwOBfF9vE4KX0Nxe"
-    "LwjcDTpsuh3qXEaZ992r1N38VDcyS6P7I6HBYN9BsNHM362zZnY27GpTw+Kwd751"
-    "CLoXFPoaMOe57dbBpXoro6Pd3BTbf/Tzr88K06yEOTDKPNj3+inbMaVigtK4PLyP"
-    "q+Topyzvx9USFgRvyuoxn0Hgb+R0A3j6SLRuyOdAi4gv7Y5oliynrSIEIAYGBgYG"
-    "BgYGBgYGBgYGBgYGBgYGBgYGBgYGBgYGBgYGrgMEAQevAwQBBLADBAEF";
-
-// kBadSessionTrailingData is a custom serialized SSL_SESSION with trailing data
-// appended.
-static const char kBadSessionTrailingData[] =
-    "MIIBdgIBAQICAwMEAsAvBCAG5Q1ndq4Yfmbeo1zwLkNRKmCXGdNgWvGT3cskV0yQ"
-    "kAQwJlrlzkAWBOWiLj/jJ76D7l+UXoizP2KI2C7I2FccqMmIfFmmkUy32nIJ0mZH"
-    "IWoJoQYCBFRDO46iBAICASykAwQBAqUDAgEUphAEDnd3dy5nb29nbGUuY29tqAcE"
-    "BXdvcmxkqQUCAwGJwKqBpwSBpBwUQvoeOk0Kg36SYTcLEkXqKwOBfF9vE4KX0Nxe"
-    "LwjcDTpsuh3qXEaZ992r1N38VDcyS6P7I6HBYN9BsNHM362zZnY27GpTw+Kwd751"
-    "CLoXFPoaMOe57dbBpXoro6Pd3BTbf/Tzr88K06yEOTDKPNj3+inbMaVigtK4PLyP"
-    "q+Topyzvx9USFgRvyuoxn0Hgb+R0A3j6SLRuyOdAi4gv7Y5oliynrSIEIAYGBgYG"
-    "BgYGBgYGBgYGBgYGBgYGBgYGBgYGBgYGBgYGrgMEAQevAwQBBLADBAEFAAAA";
-
-static bool DecodeBase64(std::vector<uint8_t> *out, const char *in) {
-  size_t len;
-  if (!EVP_DecodedLength(&len, strlen(in))) {
-    fprintf(stderr, "EVP_DecodedLength failed\n");
-    return false;
-  }
-
-  out->resize(len);
-  if (!EVP_DecodeBase64(out->data(), &len, len, (const uint8_t *)in,
-                        strlen(in))) {
-    fprintf(stderr, "EVP_DecodeBase64 failed\n");
-    return false;
-  }
-  out->resize(len);
-  return true;
-}
-
-TEST(SSLTest, SessionEncoding) {
-  for (const char *input_b64 : {
-           kOpenSSLSession,
-           kCustomSession,
-           kBoringSSLSession,
-       }) {
-    SCOPED_TRACE(std::string(input_b64));
-    // Decode the input.
-    std::vector<uint8_t> input;
-    ASSERT_TRUE(DecodeBase64(&input, input_b64));
-
-    // Verify the SSL_SESSION decodes.
-    bssl::UniquePtr<SSL_CTX> ssl_ctx(SSL_CTX_new(TLS_method()));
-    ASSERT_TRUE(ssl_ctx);
-    bssl::UniquePtr<SSL_SESSION> session(
-        SSL_SESSION_from_bytes(input.data(), input.size(), ssl_ctx.get()));
-    ASSERT_TRUE(session) << "SSL_SESSION_from_bytes failed";
-
-    // Verify the SSL_SESSION encoding round-trips.
-    size_t encoded_len;
-    bssl::UniquePtr<uint8_t> encoded;
-    uint8_t *encoded_raw;
-    ASSERT_TRUE(SSL_SESSION_to_bytes(session.get(), &encoded_raw, &encoded_len))
-        << "SSL_SESSION_to_bytes failed";
-    encoded.reset(encoded_raw);
-    EXPECT_EQ(Bytes(encoded.get(), encoded_len), Bytes(input))
-        << "SSL_SESSION_to_bytes did not round-trip";
-
-    // Verify the SSL_SESSION also decodes with the legacy API.
-    const uint8_t *cptr = input.data();
-    session.reset(d2i_SSL_SESSION(NULL, &cptr, input.size()));
-    ASSERT_TRUE(session) << "d2i_SSL_SESSION failed";
-    EXPECT_EQ(cptr, input.data() + input.size());
-
-    // Verify the SSL_SESSION encoding round-trips via the legacy API.
-    int len = i2d_SSL_SESSION(session.get(), NULL);
-    ASSERT_GT(len, 0) << "i2d_SSL_SESSION failed";
-    ASSERT_EQ(static_cast<size_t>(len), input.size())
-        << "i2d_SSL_SESSION(NULL) returned invalid length";
-
-    encoded.reset((uint8_t *)OPENSSL_malloc(input.size()));
-    ASSERT_TRUE(encoded);
-
-    uint8_t *ptr = encoded.get();
-    len = i2d_SSL_SESSION(session.get(), &ptr);
-    ASSERT_GT(len, 0) << "i2d_SSL_SESSION failed";
-    ASSERT_EQ(static_cast<size_t>(len), input.size())
-        << "i2d_SSL_SESSION(NULL) returned invalid length";
-    ASSERT_EQ(ptr, encoded.get() + input.size())
-        << "i2d_SSL_SESSION did not advance ptr correctly";
-    EXPECT_EQ(Bytes(encoded.get(), encoded_len), Bytes(input))
-        << "SSL_SESSION_to_bytes did not round-trip";
-
-    // Verify that |i2d_SSL_SESSION| works correctly when |pp| is non-NULL, but
-    // |*pp| is NULL. A newly-allocated buffer containing the result should be
-    // created. See |i2d_SAMPLE| for more details.
-    uint8_t *ptr2 = nullptr;
-    int len2 = i2d_SSL_SESSION(session.get(), &ptr2);
-    ASSERT_TRUE(ptr2);
-    ASSERT_GT(len2, 0);
-    bssl::UniquePtr<uint8_t> encoded2(ptr2);
-    EXPECT_EQ(Bytes(encoded2.get(), len2), Bytes(input))
-        << "SSL_SESSION_to_bytes did not round-trip";
-  }
-
-  for (const char *input_b64 : {
-           kBadSessionExtraField,
-           kBadSessionVersion,
-           kBadSessionTrailingData,
-       }) {
-    SCOPED_TRACE(std::string(input_b64));
-    std::vector<uint8_t> input;
-    ASSERT_TRUE(DecodeBase64(&input, input_b64));
-
-    // Verify that the SSL_SESSION fails to decode.
-    bssl::UniquePtr<SSL_CTX> ssl_ctx(SSL_CTX_new(TLS_method()));
-    ASSERT_TRUE(ssl_ctx);
-    bssl::UniquePtr<SSL_SESSION> session(
-        SSL_SESSION_from_bytes(input.data(), input.size(), ssl_ctx.get()));
-    EXPECT_FALSE(session) << "SSL_SESSION_from_bytes unexpectedly succeeded";
-    ERR_clear_error();
-  }
-}
 
 static void ExpectDefaultVersion(uint16_t min_version, uint16_t max_version,
                                  const SSL_METHOD *(*method)(void)) {
@@ -1986,129 +1031,7 @@ TEST(SSLTest, CipherProperties) {
   }
 }
 
-// CreateSessionWithTicket returns a sample |SSL_SESSION| with the specified
-// version and ticket length or nullptr on failure.
-static bssl::UniquePtr<SSL_SESSION> CreateSessionWithTicket(uint16_t version,
-                                                            size_t ticket_len) {
-  std::vector<uint8_t> der;
-  if (!DecodeBase64(&der, kOpenSSLSession)) {
-    return nullptr;
-  }
 
-  bssl::UniquePtr<SSL_CTX> ssl_ctx(SSL_CTX_new(TLS_method()));
-  if (!ssl_ctx) {
-    return nullptr;
-  }
-  // Use a garbage ticket.
-  std::vector<uint8_t> ticket(ticket_len, 'a');
-  bssl::UniquePtr<SSL_SESSION> session(
-      SSL_SESSION_from_bytes(der.data(), der.size(), ssl_ctx.get()));
-  if (!session || !SSL_SESSION_set_protocol_version(session.get(), version) ||
-      !SSL_SESSION_set_ticket(session.get(), ticket.data(), ticket.size())) {
-    return nullptr;
-  }
-  // Fix up the timeout.
-#if defined(BORINGSSL_UNSAFE_DETERMINISTIC_MODE)
-  SSL_SESSION_set_time(session.get(), 1234);
-#else
-  SSL_SESSION_set_time(session.get(), time(nullptr));
-#endif
-  return session;
-}
-
-
-
-// GetClientHelloLen creates a client SSL connection with the specified version
-// and ticket length. It returns the length of the ClientHello, not including
-// the record header, on success and zero on error.
-static size_t GetClientHelloLen(uint16_t max_version, uint16_t session_version,
-                                size_t ticket_len) {
-  bssl::UniquePtr<SSL_CTX> ctx(SSL_CTX_new(TLS_method()));
-  bssl::UniquePtr<SSL_SESSION> session =
-      CreateSessionWithTicket(session_version, ticket_len);
-  if (!ctx || !session) {
-    return 0;
-  }
-
-  // Set a one-element cipher list so the baseline ClientHello is unpadded.
-  bssl::UniquePtr<SSL> ssl(SSL_new(ctx.get()));
-  if (!ssl || !SSL_set_session(ssl.get(), session.get()) ||
-      !SSL_set_strict_cipher_list(ssl.get(), "ECDHE-RSA-AES128-GCM-SHA256") ||
-      !SSL_set_max_proto_version(ssl.get(), max_version)) {
-    return 0;
-  }
-
-  std::vector<uint8_t> client_hello;
-  if (!GetClientHello(ssl.get(), &client_hello) ||
-      client_hello.size() <= SSL3_RT_HEADER_LENGTH) {
-    return 0;
-  }
-
-  return client_hello.size() - SSL3_RT_HEADER_LENGTH;
-}
-
-TEST(SSLTest, Padding) {
-  struct PaddingVersions {
-    uint16_t max_version, session_version;
-  };
-  static const PaddingVersions kPaddingVersions[] = {
-      // Test the padding extension at TLS 1.2.
-      {TLS1_2_VERSION, TLS1_2_VERSION},
-      // Test the padding extension at TLS 1.3 with a TLS 1.2 session, so there
-      // will be no PSK binder after the padding extension.
-      {TLS1_3_VERSION, TLS1_2_VERSION},
-      // Test the padding extension at TLS 1.3 with a TLS 1.3 session, so there
-      // will be a PSK binder after the padding extension.
-      {TLS1_3_VERSION, TLS1_3_VERSION},
-
-  };
-
-  struct PaddingTest {
-    size_t input_len, padded_len;
-  };
-  static const PaddingTest kPaddingTests[] = {
-      // ClientHellos of length below 0x100 do not require padding.
-      {0xfe, 0xfe},
-      {0xff, 0xff},
-      // ClientHellos of length 0x100 through 0x1fb are padded up to 0x200.
-      {0x100, 0x200},
-      {0x123, 0x200},
-      {0x1fb, 0x200},
-      // ClientHellos of length 0x1fc through 0x1ff get padded beyond 0x200. The
-      // padding extension takes a minimum of four bytes plus one required
-      // content
-      // byte. (To work around yet more server bugs, we avoid empty final
-      // extensions.)
-      {0x1fc, 0x201},
-      {0x1fd, 0x202},
-      {0x1fe, 0x203},
-      {0x1ff, 0x204},
-      // Finally, larger ClientHellos need no padding.
-      {0x200, 0x200},
-      {0x201, 0x201},
-  };
-
-  for (const PaddingVersions &versions : kPaddingVersions) {
-    SCOPED_TRACE(versions.max_version);
-    SCOPED_TRACE(versions.session_version);
-
-    // Sample a baseline length.
-    size_t base_len =
-        GetClientHelloLen(versions.max_version, versions.session_version, 1);
-    ASSERT_NE(base_len, 0u) << "Baseline length could not be sampled";
-
-    for (const PaddingTest &test : kPaddingTests) {
-      SCOPED_TRACE(test.input_len);
-      ASSERT_LE(base_len, test.input_len) << "Baseline ClientHello too long";
-
-      size_t padded_len =
-          GetClientHelloLen(versions.max_version, versions.session_version,
-                            1 + test.input_len - base_len);
-      EXPECT_EQ(padded_len, test.padded_len)
-          << "ClientHello was not padded to expected length";
-    }
-  }
-}
 
 
 
@@ -2122,32 +1045,6 @@ static bssl::UniquePtr<SSL_CTX> CreateContextWithCertificate(
     return nullptr;
   }
   return ctx;
-}
-
-
-
-
-
-static bssl::UniquePtr<X509> GetED25519TestCertificate() {
-  static const char kCertPEM[] =
-      "-----BEGIN CERTIFICATE-----\n"
-      "MIIBRDCB9wIUKI+32tShPulvafJa3xZvj29Z9xgwBQYDK2VwMEUxCzAJBgNVBAYT\n"
-      "AkFVMRMwEQYDVQQIDApTb21lLVN0YXRlMSEwHwYDVQQKDBhJbnRlcm5ldCBXaWRn\n"
-      "aXRzIFB0eSBMdGQwHhcNMjMwNzE4MTg0NzU4WhcNMjMwNzE5MTg0NzU4WjBFMQsw\n"
-      "CQYDVQQGEwJBVTETMBEGA1UECAwKU29tZS1TdGF0ZTEhMB8GA1UECgwYSW50ZXJu\n"
-      "ZXQgV2lkZ2l0cyBQdHkgTHRkMCowBQYDK2VwAyEAprAzqgxux8R4ZXaxn5mM/5E9\n"
-      "0RNE59r47BJikdOoeUwwBQYDK2VwA0EAMELt0XRGFYo4qkWwOsoSYcdGYqlxVlf9\n"
-      "AhTPaJ6SSzjv3n4r60wfe8Z2OPn415tcj2IIm42T64itI4OAX0aTCg==\n"
-      "-----END CERTIFICATE-----\n";
-  return CertFromPEM(kCertPEM);
-}
-
-static bssl::UniquePtr<EVP_PKEY> GetED25519TestKey() {
-  static const char kKeyPEM[] =
-      "-----BEGIN PRIVATE KEY-----\n"
-      "MC4CAQAwBQYDK2VwBCIEIGPkz4xAobc5gtRidkHl+fxNLHfiWo3efRG2G8Z617yk\n"
-      "-----END PRIVATE KEY-----\n";
-  return KeyFromPEM(kKeyPEM);
 }
 
 // Correct ID and name
@@ -4679,63 +3576,6 @@ TEST(SSLTest, SignatureAlgorithmProperties) {
   EXPECT_TRUE(SSL_is_signature_algorithm_rsa_pss(SSL_SIGN_RSA_PSS_RSAE_SHA384));
 }
 
-static int XORCompressFunc(SSL *ssl, CBB *out, const uint8_t *in,
-                           size_t in_len) {
-  for (size_t i = 0; i < in_len; i++) {
-    if (!CBB_add_u8(out, in[i] ^ 0x55)) {
-      return 0;
-    }
-  }
-
-  SSL_set_app_data(ssl, XORCompressFunc);
-
-  return 1;
-}
-
-static int XORDecompressFunc(SSL *ssl, CRYPTO_BUFFER **out,
-                             size_t uncompressed_len, const uint8_t *in,
-                             size_t in_len) {
-  if (in_len != uncompressed_len) {
-    return 0;
-  }
-
-  uint8_t *data;
-  *out = CRYPTO_BUFFER_alloc(&data, uncompressed_len);
-  if (*out == nullptr) {
-    return 0;
-  }
-
-  for (size_t i = 0; i < in_len; i++) {
-    data[i] = in[i] ^ 0x55;
-  }
-
-  SSL_set_app_data(ssl, XORDecompressFunc);
-
-  return 1;
-}
-
-TEST(SSLTest, CertCompression) {
-  bssl::UniquePtr<SSL_CTX> client_ctx(SSL_CTX_new(TLS_method()));
-  bssl::UniquePtr<SSL_CTX> server_ctx(
-      CreateContextWithTestCertificate(TLS_method()));
-  ASSERT_TRUE(client_ctx);
-  ASSERT_TRUE(server_ctx);
-
-  ASSERT_TRUE(SSL_CTX_set_max_proto_version(client_ctx.get(), TLS1_3_VERSION));
-  ASSERT_TRUE(SSL_CTX_set_max_proto_version(server_ctx.get(), TLS1_3_VERSION));
-  ASSERT_TRUE(SSL_CTX_add_cert_compression_alg(
-      client_ctx.get(), 0x1234, XORCompressFunc, XORDecompressFunc));
-  ASSERT_TRUE(SSL_CTX_add_cert_compression_alg(
-      server_ctx.get(), 0x1234, XORCompressFunc, XORDecompressFunc));
-
-  bssl::UniquePtr<SSL> client, server;
-  ASSERT_TRUE(ConnectClientAndServer(&client, &server, client_ctx.get(),
-                                     server_ctx.get()));
-
-  EXPECT_TRUE(SSL_get_app_data(client.get()) == XORDecompressFunc);
-  EXPECT_TRUE(SSL_get_app_data(server.get()) == XORCompressFunc);
-}
-
 
 
 TEST(SSLTest, HandoffDeclined) {
@@ -4999,187 +3839,6 @@ TEST(SSLTest, ApplyHandoffRemovesUnsupportedCurves) {
   EXPECT_EQ(1u, server->config->supported_group_list.size());
 }
 
-struct EncodeDecodeKATTestParam {
-  const char *input;
-  const char *output;
-};
-
-static const EncodeDecodeKATTestParam kEncodeDecodeKATs[] = {
-    // V1 input round-trips as V2 output
-    {"308201173082011302010102020303020240003081fa0201010408000000000000000104"
-     "0800000000000000010420000004d29e62f41ded4bb33d0faa6ffada380e2c489dfbfb44"
-     "4f574e475244010420cf3926d1ec5a562a642935a8050222b0aed93ffd9d1cac682274d9"
-     "42e99e42a604020000020100020103040cb9b409f5129440622f87f84402010c040c1f49"
-     "e2e989c66a263e9c227502010c020100020100020100a05b3059020101020203030402cc"
-     "a80400043085668dcf9f0921094ebd7f91bf2a8c60d276e4c279fd85a989402f67868232"
-     "4fd8098dc19d900b856d0a77e048e3ced2a104020204d2a20402021c20a4020400b10301"
-     "01ffb20302011da206040474657374a7030101ff020108020100a0030101ff",
-     "308201803082017c02010102020303020240003082016202010204080000000000000001"
-     "040800000000000000010420000004d29e62f41ded4bb33d0faa6ffada380e2c489dfbfb"
-     "444f574e475244010420cf3926d1ec5a562a642935a8050222b0aed93ffd9d1cac682274"
-     "d942e99e42a6040200000201000201030440b9b409f5129440622f87f84402010c040c1f"
-     "49e2e989c66a263e9c227502010c020100020100020100a05b3059020101020203030402"
-     "cca80400043085668dcf02010c04401f49e2e989c66a263e9c227502010c020100020100"
-     "020100a05b3059020101020203030402cca80400043085668dcf9f0921094ebd7f91bf2a"
-     "8c60d276e4c27902010c020100020100020100a05b3059020101020203030402cca80400"
-     "043085668dcf9f0921094ebd7f91bf2a8c60d276e4c279fd85a989402f678682324fd809"
-     "8dc19d900b856d0a77e048e3ced2a104020204d2a20402021c20a4020400b1030101ffb2"
-     "0302011da206040474657374a7030101ff020108020100a0030101ff"},
-    // In runner.go, the test case "Basic-Server-TLS-Sync-SSL_Transfer" is used
-    // to generate below bytes by adding print statement on the output of
-    // |SSL_to_bytes| in bssl_shim.cc.
-    // We've bumped the buffer size in the |previous_client/server_finished|
-    // fields. This verifies that the original size is parsable and reencoded
-    // with the new size.
-    {"308201173082011302010102020303020240003081fa0201020408000000000000000104"
-     "0800000000000000010420000004d29e62f41ded4bb33d0faa6ffada380e2c489dfbfb44"
-     "4f574e475244010420cf3926d1ec5a562a642935a8050222b0aed93ffd9d1cac682274d9"
-     "42e99e42a604020000020100020103040cb9b409f5129440622f87f84402010c040c1f49"
-     "e2e989c66a263e9c227502010c020100020100020100a05b3059020101020203030402cc"
-     "a80400043085668dcf9f0921094ebd7f91bf2a8c60d276e4c279fd85a989402f67868232"
-     "4fd8098dc19d900b856d0a77e048e3ced2a104020204d2a20402021c20a4020400b10301"
-     "01ffb20302011da206040474657374a7030101ff020108020100a0030101ff",
-     "308201803082017c02010102020303020240003082016202010204080000000000000001"
-     "040800000000000000010420000004d29e62f41ded4bb33d0faa6ffada380e2c489dfbfb"
-     "444f574e475244010420cf3926d1ec5a562a642935a8050222b0aed93ffd9d1cac682274"
-     "d942e99e42a6040200000201000201030440b9b409f5129440622f87f84402010c040c1f"
-     "49e2e989c66a263e9c227502010c020100020100020100a05b3059020101020203030402"
-     "cca80400043085668dcf02010c04401f49e2e989c66a263e9c227502010c020100020100"
-     "020100a05b3059020101020203030402cca80400043085668dcf9f0921094ebd7f91bf2a"
-     "8c60d276e4c27902010c020100020100020100a05b3059020101020203030402cca80400"
-     "043085668dcf9f0921094ebd7f91bf2a8c60d276e4c279fd85a989402f678682324fd809"
-     "8dc19d900b856d0a77e048e3ced2a104020204d2a20402021c20a4020400b1030101ffb2"
-     "0302011da206040474657374a7030101ff020108020100a0030101ff"},
-     // In runner.go, the test case
-     // "TLS-TLS13-AES_128_GCM_SHA256-server-SSL_Transfer" is used to generate
-     // below bytes by adding print statement on the output of |SSL_to_bytes| in
-     // bssl_shim.cc.
-     // We've bumped the buffer size in the |previous_client/server_finished|
-     // fields. This verifies that the original size is parsable and reencoded
-     // with the new size.
-    {"308203883082038402010102020304020240003082036a020102040800000000000000000"
-      "408000000000000000004206beca5c14aff6b92757545948b883c6c175327814bedcf38a6"
-      "b2e4c43bc02d180420a32aee5b7705a19e4bb2b47f4918199c76cee7245f1311bc4ba3888"
-      "3d33f236a04020000020100020101040c000000000000000000000000020100040c000000"
-      "000000000000000000020100020100020100020100a04e304c02010102020304040213010"
-      "40004200b66320d38c8fa1b0dfe9e37fcf2bf0bafb43077fa31ed2f1220dd245cef4c4da1"
-      "04020204d2a205020302a300a4020400b20302011db9050203093a80a206040474657374a"
-      "b03020100ac03010100ad03010100ae03010100af03020100b032043034c0893be938bade"
-      "e7029ca3cfea4c821dde48e03f0d07641cba33b247bc161c0000000000000000000000000"
-      "0000000b103020120b232043094b319ed2f41ee11aa73e141a238e5724c04f2aa8298c16b"
-      "43c910c40cc98d1500000000000000000000000000000000b303020120b432043015a178c"
-      "e69c0110ad36da8d58ca8428d9615ff07fc6a4e1bbab026c1bb0c02180000000000000000"
-      "0000000000000000b503020120b88201700482016c040000b20002a30056355452010000a"
-      "027abfd1f1aa28cee6e8e2396112e8285f150768898158dbce97a1aef0a63fa6dda1002a4"
-      "d75942a3739c11e4b25827f529ab59d22e34e0cf0b59b9336eb60edbb1f686c072ab33c30"
-      "e784f876da5b4c7fddd67f4a2ffa995f8c9ccf2128200ae9668d626866b1b7c6bb111867a"
-      "87ed2a96122736595374f8fe5343e6ca492b278b67b1571423f2c1bcb673922e9044e9094"
-      "9975ff72ab4a0eb659d8de664cac600042a2a0000040000b20002a3009e8c6738010100a0"
-      "27abfd1f1aa28cee6e8e2396112e82851f15c84668b2f1d717681d1a3c6d2ea52d3401d31"
-      "10a04498246480b96a7e5b3c39ea6cef3a2a86b81896f1621950472d858d18796c97e8320"
-      "4daf94c1f30dfe763cd282fbee718a679dca8bff3cc8e11724062232e573bcf0252dc4d39"
-      "0baa2b7f49a164b46d2d685e9fe826465cc135130f3e2e47838658af57173f864070fdce2"
-      "41be58ecbd60d18128dfa28f4b1a00042a2a0000ba2330210201010204030013013016020"
-      "101020117040e300c0201010201000201000101ffbb233021020101020403001301301602"
-      "0101020117040e300c0201010201000201000101ff020108020100a0030101ff",
-      "308203f0308203ec0201010202030402024000308203d202010204080000000000000000"
-      "0408000000000000000004206beca5c14aff6b92757545948b883c6c175327814bedcf38"
-      "a6b2e4c43bc02d180420a32aee5b7705a19e4bb2b47f4918199c76cee7245f1311bc4ba3"
-      "8883d33f236a040200000201000201010440000000000000000000000000020100040c00"
-      "0000000000000000000000020100020100020100020100a04e304c020101020203040402"
-      "1301040004200b66320d0201000440000000000000000000000000020100020100020100"
-      "020100a04e304c0201010202030404021301040004200b66320d38c8fa1b0dfe9e37fcf2"
-      "bf0bafb43077fa020100020100020100020100a04e304c02010102020304040213010400"
-      "04200b66320d38c8fa1b0dfe9e37fcf2bf0bafb43077fa31ed2f1220dd245cef4c4da104"
-      "020204d2a205020302a300a4020400b20302011db9050203093a80a206040474657374ab"
-      "03020100ac03010100ad03010100ae03010100af03020100b032043034c0893be938bade"
-      "e7029ca3cfea4c821dde48e03f0d07641cba33b247bc161c000000000000000000000000"
-      "00000000b103020120b232043094b319ed2f41ee11aa73e141a238e5724c04f2aa8298c1"
-      "6b43c910c40cc98d1500000000000000000000000000000000b303020120b432043015a1"
-      "78ce69c0110ad36da8d58ca8428d9615ff07fc6a4e1bbab026c1bb0c0218000000000000"
-      "00000000000000000000b503020120b88201700482016c040000b20002a3005635545201"
-      "0000a027abfd1f1aa28cee6e8e2396112e8285f150768898158dbce97a1aef0a63fa6dda"
-      "1002a4d75942a3739c11e4b25827f529ab59d22e34e0cf0b59b9336eb60edbb1f686c072"
-      "ab33c30e784f876da5b4c7fddd67f4a2ffa995f8c9ccf2128200ae9668d626866b1b7c6b"
-      "b111867a87ed2a96122736595374f8fe5343e6ca492b278b67b1571423f2c1bcb673922e"
-      "9044e90949975ff72ab4a0eb659d8de664cac600042a2a0000040000b20002a3009e8c67"
-      "38010100a027abfd1f1aa28cee6e8e2396112e82851f15c84668b2f1d717681d1a3c6d2e"
-      "a52d3401d3110a04498246480b96a7e5b3c39ea6cef3a2a86b81896f1621950472d858d1"
-      "8796c97e83204daf94c1f30dfe763cd282fbee718a679dca8bff3cc8e11724062232e573"
-      "bcf0252dc4d390baa2b7f49a164b46d2d685e9fe826465cc135130f3e2e47838658af571"
-      "73f864070fdce241be58ecbd60d18128dfa28f4b1a00042a2a0000ba2330210201010204"
-      "030013013016020101020117040e300c0201010201000201000101ffbb23302102010102"
-      "04030013013016020101020117040e300c0201010201000201000101ff020108020100a0"
-      "030101ff"},
-    // In runner.go, the test case
-    // "TLS-ECH-Server-Cipher-HKDF-SHA256-AES-256-GCM-SSL_Transfer" is used
-    // to generate below bytes by adding print statement on the output of
-    // |SSL_to_bytes| in bssl_shim.cc.
-    {"308203e3308203df0201010202030402024000308203c502010204080000000000000000"
-     "04080000000000000000042028431b914ffdb44ea92ca53d5734976c6a16f141d44f180b"
-     "0816a5cb2b8e79030420bdaf544fa82d833d58c92213e44e850cc0b8147699b0b410d4aa"
-     "2a277030f3220402000002010002010104409e155007d04cd03cf4d8a95ce244dc978a87"
-     "e1808f0f6c6acb51ad7bf8063ae000000000000000000000000000000000000000000000"
-     "0000000000000000000002012004406680e8c36429d465ea520ae74a2062a5e07c39f34b"
-     "688024ae2edfab2898670700000000000000000000000000000000000000000000000000"
-     "00000000000000020120020100020100020100a04e304c02010102020304040213030400"
-     "0420df74ecd172087ad53083d505145ec4f6cf0ec5ed64b67ba526d55c918a0f8936a104"
-     "020204d2a205020302a300a4020400b20302011db9050203093a80a210040e7365637265"
-     "742e6578616d706c65ab03020100ac03010100ad03010100ae03010100af03020100b032"
-     "0430c40f9f95646fa700d58934e79c36b84ba3502d33df04248d56cded3444927e300000"
-     "0000000000000000000000000000b103020120b23204307a1a99bf276b5e5be57dd68968"
-     "411594e77b1a48cf2c03cc5c143985aa40b32e00000000000000000000000000000000b3"
-     "03020120b4320430cbf50af88bc5a610910139172a468663675882caacaf176aa961b12a"
-     "38a0df2a00000000000000000000000000000000b503020120b703020101b88201700482"
-     "016c040000b20002a300bbccf972010000a041e0b13ecd71dfb3d9e3cb451e37cfde8197"
-     "3a1b73106b6669b53475781f0203a3f32f45cef7742cf0efb86d850081254f20d3b6bd83"
-     "30bc70331464905bcd99383c33e42c7d34bfeb47b387bf43b5c796daa4581f8b0043b7eb"
-     "216911f8eebaf1e8bd5d05277943d5a319cc03d9555e414990099f56ee887145f34e8bff"
-     "27f06d1865aa64d548a22208318566959a097c080fa3e5e0d4b1d933132ef32929950004"
-     "5a5a0000040000b20002a3002ecba343010100a041e0b13ecd71dfb3d9e3cb451e37cfde"
-     "289f90201519fb0dff08aa9e14a9f4ee1434edce481e49d22f061529bb4d230258f3dac8"
-     "86c2c1100bee2ccc7be889a90b417270c30b3b770558ef6f3c444ddefd08e673f788931d"
-     "86542c4a1e7ec44b0957bb315c17851bd8498b1d1131a79e19c66463e0566985ef55deb5"
-     "48fe370058ba83566278d01b3a565075b8ef2a82bea17ae95fa91b7b3ffa611a7d8a6331"
-     "00045a5a0000ba15301302010102040300130330080201010201050400bb153013020101"
-     "02040300130330080201010201050400020108020100a0030101ff", nullptr}
-};
-
-class EncodeDecodeKATTest
-    : public testing::TestWithParam<EncodeDecodeKATTestParam> {};
-
-INSTANTIATE_TEST_SUITE_P(EncodeAndDecodeKATTests, EncodeDecodeKATTest,
-                         testing::ValuesIn(kEncodeDecodeKATs));
-
-TEST_P(EncodeDecodeKATTest, RoundTrips) {
-  std::string input(GetParam().input);
-  std::string output;
-  if (GetParam().output) {
-    output = std::string(GetParam().output);
-  } else {
-    output = std::string(GetParam().input);
-  }
-
-  std::vector<uint8_t> input_bytes;
-  ASSERT_TRUE(DecodeHex(&input_bytes, input));
-  std::vector<uint8_t> output_bytes;
-  ASSERT_TRUE(DecodeHex(&output_bytes, output));
-
-  bssl::UniquePtr<SSL_CTX> server_ctx(SSL_CTX_new(TLS_method()));
-  // Check the bytes are decoded successfully.
-  bssl::UniquePtr<SSL> ssl(
-      SSL_from_bytes(input_bytes.data(), input_bytes.size(), server_ctx.get()));
-  ASSERT_TRUE(ssl);
-  // Check the ssl can be encoded successfully.
-  size_t encoded_len;
-  uint8_t *encoded;
-  ASSERT_TRUE(SSL_to_bytes(ssl.get(), &encoded, &encoded_len));
-  bssl::UniquePtr<uint8_t> encoded_ptr;
-  encoded_ptr.reset(encoded);
-  // Check the encoded bytes are the same as the test input.
-  ASSERT_EQ(output_bytes.size(), encoded_len);
-  ASSERT_EQ(OPENSSL_memcmp(output_bytes.data(), encoded, encoded_len), 0);
-}
 
 TEST(SSLTest, ZeroSizedWriteFlushesHandshakeMessages) {
   // If there are pending handshake messages, an |SSL_write| of zero bytes
@@ -6415,536 +5074,6 @@ TEST(SSLTest, ErrorSyscallAfterCloseNotify) {
   EXPECT_EQ(SSL_get_error(client.get(), ret), SSL_ERROR_SYSCALL);
   EXPECT_TRUE(write_failed);
   write_failed = false;
-}
-
-// Test that failures are supressed on (potentially)
-// transient empty reads.
-TEST(SSLTest, IntermittentEmptyRead) {
-  bssl::UniquePtr<SSL_CTX> client_ctx(SSL_CTX_new(TLS_method()));
-  bssl::UniquePtr<SSL_CTX> server_ctx =
-      CreateContextWithTestCertificate(TLS_method());
-  ASSERT_TRUE(client_ctx);
-  ASSERT_TRUE(server_ctx);
-  bssl::UniquePtr<SSL> client, server;
-  ASSERT_TRUE(ConnectClientAndServer(&client, &server, client_ctx.get(),
-                                     server_ctx.get()));
-
-  // Create a fake read BIO that returns 0 on read to simulate empty read
-  bssl::UniquePtr<BIO_METHOD> method(BIO_meth_new(0, nullptr));
-  ASSERT_TRUE(method);
-  ASSERT_TRUE(BIO_meth_set_create(method.get(), [](BIO *b) -> int {
-    BIO_set_init(b, 1);
-    return 1;
-  }));
-  ASSERT_TRUE(BIO_meth_set_read(method.get(), [](BIO *, char *, int) -> int {
-    return 0;
-  }));
-  bssl::UniquePtr<BIO> rbio_empty(BIO_new(method.get()));
-  ASSERT_TRUE(rbio_empty);
-  BIO_set_flags(rbio_empty.get(), BIO_FLAGS_READ);
-
-  // Save off client rbio and use empty read BIO
-  bssl::UniquePtr<BIO> client_rbio(SSL_get_rbio(client.get()));
-  ASSERT_TRUE(client_rbio);
-  // Up-ref |client_rbio| as SSL_CTX dtor will also attempt to free it
-  ASSERT_TRUE(BIO_up_ref(client_rbio.get()));
-  SSL_set0_rbio(client.get(), rbio_empty.release());
-
-  // Server writes some data to the client
-  const uint8_t write_data[] = {1, 2, 3};
-  int ret = SSL_write(server.get(), write_data, (int) sizeof(write_data));
-  EXPECT_EQ(ret, (int) sizeof(write_data));
-  EXPECT_EQ(SSL_get_error(server.get(), ret), SSL_ERROR_NONE);
-
-  uint8_t read_data[] = {0, 0, 0};
-  ret = SSL_read(client.get(), read_data, sizeof(read_data));
-  EXPECT_EQ(ret, 0);
-  // On empty read, client should still want a read so caller will retry.
-  // This would have returned |SSL_ERROR_SYSCALL| in OpenSSL 1.0.2.
-  EXPECT_EQ(SSL_get_error(client.get(), ret), SSL_ERROR_WANT_READ);
-
-  // Reset client rbio, read should succeed
-  SSL_set0_rbio(client.get(), client_rbio.release());
-  ret = SSL_read(client.get(), read_data, sizeof(read_data));
-  EXPECT_EQ(ret, (int) sizeof(write_data));
-  EXPECT_EQ(OPENSSL_memcmp(read_data, write_data, sizeof(write_data)), 0);
-  EXPECT_EQ(SSL_get_error(client.get(), ret), SSL_ERROR_NONE);
-
-  // Subsequent attempts to read should fail
-  ret = SSL_read(client.get(), read_data, sizeof(read_data));
-  EXPECT_LT(ret, 0);
-  EXPECT_EQ(SSL_get_error(client.get(), ret), SSL_ERROR_WANT_READ);
-}
-
-// Test that |SSL_shutdown|, when quiet shutdown is enabled, simulates receiving
-// a close_notify, down to |SSL_read| reporting |SSL_ERROR_ZERO_RETURN|.
-TEST(SSLTest, QuietShutdown) {
-  bssl::UniquePtr<SSL_CTX> client_ctx(SSL_CTX_new(TLS_method()));
-  bssl::UniquePtr<SSL_CTX> server_ctx =
-      CreateContextWithTestCertificate(TLS_method());
-  ASSERT_TRUE(client_ctx);
-  ASSERT_TRUE(server_ctx);
-  SSL_CTX_set_quiet_shutdown(server_ctx.get(), 1);
-  bssl::UniquePtr<SSL> client, server;
-  ASSERT_TRUE(ConnectClientAndServer(&client, &server, client_ctx.get(),
-                                     server_ctx.get()));
-
-  // Quiet shutdown is enabled, so |SSL_shutdown| on the server should
-  // immediately return that bidirectional shutdown "completed".
-  EXPECT_EQ(SSL_shutdown(server.get()), 1);
-
-  // Shut down writes so the client gets an EOF.
-  EXPECT_TRUE(BIO_shutdown_wr(SSL_get_wbio(server.get())));
-
-  // Confirm no close notify was actually sent. Client reads should report a
-  // transport EOF, not a close_notify. (Both have zero return, but
-  // |SSL_get_error| is different.)
-  char buf[1];
-  int ret = SSL_read(client.get(), buf, sizeof(buf));
-  EXPECT_EQ(ret, 0);
-  EXPECT_EQ(SSL_get_error(client.get(), ret), SSL_ERROR_SYSCALL);
-
-  // The server believes bidirectional shutdown completed, so reads should
-  // replay the (simulated) close_notify.
-  ret = SSL_read(server.get(), buf, sizeof(buf));
-  EXPECT_EQ(ret, 0);
-  EXPECT_EQ(SSL_get_error(server.get(), ret), SSL_ERROR_ZERO_RETURN);
-}
-
-TEST(SSLTest, InvalidSignatureAlgorithm) {
-  bssl::UniquePtr<SSL_CTX> ctx(SSL_CTX_new(TLS_method()));
-  ASSERT_TRUE(ctx);
-
-  static const uint16_t kInvalidPrefs[] = {1234};
-  EXPECT_FALSE(SSL_CTX_set_signing_algorithm_prefs(
-      ctx.get(), kInvalidPrefs, OPENSSL_ARRAY_SIZE(kInvalidPrefs)));
-  EXPECT_FALSE(SSL_CTX_set_verify_algorithm_prefs(
-      ctx.get(), kInvalidPrefs, OPENSSL_ARRAY_SIZE(kInvalidPrefs)));
-
-  static const uint16_t kDuplicatePrefs[] = {SSL_SIGN_RSA_PKCS1_SHA256,
-                                             SSL_SIGN_RSA_PKCS1_SHA256};
-  EXPECT_FALSE(SSL_CTX_set_signing_algorithm_prefs(
-      ctx.get(), kDuplicatePrefs, OPENSSL_ARRAY_SIZE(kDuplicatePrefs)));
-  EXPECT_FALSE(SSL_CTX_set_verify_algorithm_prefs(
-      ctx.get(), kDuplicatePrefs, OPENSSL_ARRAY_SIZE(kDuplicatePrefs)));
-}
-
-TEST(SSLTest, ErrorStrings) {
-  int warning_value = SSL3_AD_CLOSE_NOTIFY | (SSL3_AL_WARNING << 8);
-  int fatal_value = SSL3_AD_UNEXPECTED_MESSAGE | (SSL3_AL_FATAL << 8);
-  int unknown_value = 99999;
-
-  EXPECT_EQ(Bytes(SSL_alert_desc_string(warning_value)), Bytes("CN"));
-  EXPECT_EQ(Bytes(SSL_alert_desc_string(fatal_value)), Bytes("UM"));
-  EXPECT_EQ(Bytes(SSL_alert_desc_string(unknown_value)), Bytes("UK"));
-
-  EXPECT_EQ(Bytes(SSL_alert_type_string(warning_value)), Bytes("W"));
-  EXPECT_EQ(Bytes(SSL_alert_type_string(fatal_value)), Bytes("F"));
-  EXPECT_EQ(Bytes(SSL_alert_type_string(unknown_value)), Bytes("U"));
-}
-
-TEST(SSLTest, NameLists) {
-  struct {
-    size_t (*func)(const char **, size_t);
-    std::vector<std::string> expected;
-  } kTests[] = {
-      {SSL_get_all_version_names, {"TLSv1.3", "DTLSv1.2", "unknown"}},
-      {SSL_get_all_standard_cipher_names,
-       {"TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256", "TLS_AES_128_GCM_SHA256"}},
-      {SSL_get_all_cipher_names,
-       {"ECDHE-ECDSA-AES128-GCM-SHA256", "TLS_AES_128_GCM_SHA256", "(NONE)"}},
-      {SSL_get_all_group_names, {"P-256", "X25519"}},
-      {SSL_get_all_signature_algorithm_names,
-       {"rsa_pkcs1_sha256", "ecdsa_secp256r1_sha256", "ecdsa_sha256"}},
-  };
-  for (const auto &t : kTests) {
-    size_t num = t.func(nullptr, 0);
-    EXPECT_GT(num, 0u);
-
-    std::vector<const char*> list(num);
-    EXPECT_EQ(num, t.func(list.data(), list.size()));
-
-    // Check the expected values are in the list.
-    for (const auto &s : t.expected) {
-      EXPECT_NE(list.end(), std::find(list.begin(), list.end(), s))
-          << "Could not find " << s;
-    }
-
-    // Passing in a larger buffer should leave excess space alone.
-    std::vector<const char *> list2(num + 1, "placeholder");
-    EXPECT_EQ(num, t.func(list2.data(), list2.size()));
-    for (size_t i = 0; i < num; i++) {
-      EXPECT_STREQ(list[i], list2[i]);
-    }
-    EXPECT_STREQ(list2.back(), "placeholder");
-
-    // Passing in a shorter buffer should truncate the list.
-    for (size_t l = 0; l < num; l++) {
-      SCOPED_TRACE(l);
-      list2.resize(l);
-      EXPECT_EQ(num, t.func(list2.data(), list2.size()));
-      for (size_t i = 0; i < l; i++) {
-        EXPECT_STREQ(list[i], list2[i]);
-      }
-    }
-  }
-}
-
-
-TEST(SSLTest, SessionPrint) {
- static const std::array<std::string, 15> kExpectedTLS13{
-      {"SSL-Session:", "    Protocol  :", "    Cipher    : ",
-       "    Session-ID: ", "    Session-ID-ctx:", "    Resumption PSK:",
-       "    PSK identity:", "    TLS session ticket lifetime hint:",
-       "    TLS session ticket:", "    61",
-       "    Start Time:", "    Timeout   :", "    Verify return code:",
-       "    Extended master secret:", "    Max Early Data:"}};
-
-  static const std::array<std::string, 14> kExpectedTLS12{
-      {"SSL-Session:", "    Protocol  :", "    Cipher    : ",
-       "    Session-ID: ", "    Session-ID-ctx:", "    Master-Key:",
-       "    PSK identity:", "    TLS session ticket lifetime hint:",
-       "    TLS session ticket:", "    61",
-       "    Start Time:", "    Timeout   :", "    Verify return code:",
-       "    Extended master secret:"}};
-
-  bssl::UniquePtr<SSL_SESSION> session(
-      CreateSessionWithTicket(TLS1_3_VERSION, 10));
-  bssl::UniquePtr<BIO> bio(BIO_new(BIO_s_mem()));
-  EXPECT_TRUE(SSL_SESSION_print(bio.get(), session.get()));
-  const uint8_t *out;
-  size_t outlen;
-  ASSERT_TRUE(BIO_mem_contents(bio.get(), &out, &outlen));
-
-  // Iterate through |kExpectedTLS13| and verify that |SSL_SESSION_print| has
-  // the expected format.
-  std::istringstream iss_tls13((std::string((char *)out, outlen)));
-  std::string line;
-  for (const auto &expected : kExpectedTLS13) {
-    std::getline(iss_tls13, line);
-    EXPECT_EQ(line.substr(0, expected.length()), expected);
-  }
-
-  session = CreateSessionWithTicket(TLS1_2_VERSION, 10);
-  bio.reset(BIO_new(BIO_s_mem()));
-  EXPECT_TRUE(SSL_SESSION_print(bio.get(), session.get()));
-  ASSERT_TRUE(BIO_mem_contents(bio.get(), &out, &outlen));
-  // Iterate through |kExpectedTLS12| and verify that |SSL_SESSION_print| has
-  // the expected format.
-  std::istringstream iss_tls12((std::string((char *)out, outlen)));
-  for (const auto &expected : kExpectedTLS12) {
-    std::getline(iss_tls12, line);
-    EXPECT_EQ(line.substr(0, expected.length()), expected);
-  }
-}
-
-class PerformHybridHandshakeTest : public testing::TestWithParam<HybridHandshakeTest> {};
-INSTANTIATE_TEST_SUITE_P(PerformHybridHandshakeTests, PerformHybridHandshakeTest, testing::ValuesIn(kHybridHandshakeTests));
-
-// This test runs through an overall handshake flow for all of the cases
-// defined in kHybridHandshakeTests. This test runs through both positive and
-// negative cases; refer to the comments inline in kHybridHandshakeTests for
-// specifics about each test case.
-TEST_P(PerformHybridHandshakeTest, PerformHybridHandshake) {
-  HybridHandshakeTest t = GetParam();
-  // Set up client and server with test case parameters
-  bssl::UniquePtr<SSL_CTX> client_ctx(SSL_CTX_new(TLS_method()));
-  ASSERT_TRUE(client_ctx);
-  ASSERT_TRUE(SSL_CTX_set1_curves_list(client_ctx.get(), t.client_rule));
-  ASSERT_TRUE(
-      SSL_CTX_set_max_proto_version(client_ctx.get(), t.client_version));
-
-  bssl::UniquePtr<SSL_CTX> server_ctx =
-      CreateContextWithTestCertificate(TLS_method());
-  ASSERT_TRUE(server_ctx);
-  ASSERT_TRUE(SSL_CTX_set1_curves_list(server_ctx.get(), t.server_rule));
-  ASSERT_TRUE(
-      SSL_CTX_set_max_proto_version(server_ctx.get(), t.server_version));
-
-  bssl::UniquePtr<SSL> client, server;
-  ASSERT_TRUE(CreateClientAndServer(&client, &server, client_ctx.get(),
-                                    server_ctx.get()));
-
-  if (t.expected_group != 0) {
-    // In this case, assert that the handshake completes as expected.
-    ASSERT_TRUE(CompleteHandshakes(client.get(), server.get()));
-
-    SSL_SESSION *client_session = SSL_get_session(client.get());
-    ASSERT_TRUE(client_session);
-    EXPECT_EQ(t.expected_group, client_session->group_id);
-    EXPECT_EQ(t.is_hrr_expected, SSL_used_hello_retry_request(client.get()));
-    EXPECT_EQ(std::min(t.client_version, t.server_version),
-              client_session->ssl_version);
-
-    SSL_SESSION *server_session = SSL_get_session(server.get());
-    ASSERT_TRUE(server_session);
-    EXPECT_EQ(t.expected_group, server_session->group_id);
-    EXPECT_EQ(t.is_hrr_expected, SSL_used_hello_retry_request(server.get()));
-    EXPECT_EQ(std::min(t.client_version, t.server_version),
-              server_session->ssl_version);
-  } else {
-    // In this case, we expect the handshake to fail because client and
-    // server configurations are not compatible.
-    ASSERT_FALSE(CompleteHandshakes(client.get(), server.get()));
-
-    ASSERT_FALSE(client.get()->s3->initial_handshake_complete);
-    EXPECT_EQ(t.is_hrr_expected, SSL_used_hello_retry_request(client.get()));
-
-    ASSERT_FALSE(server.get()->s3->initial_handshake_complete);
-    EXPECT_EQ(t.is_hrr_expected, SSL_used_hello_retry_request(server.get()));
-  }
-}
-
-TEST(SSLTest, SSLFileTests) {
-#if defined(OPENSSL_ANDROID)
-  // On Android, when running from an APK, temporary file creations do not work.
-  // See b/36991167#comment8.
-  GTEST_SKIP();
-#endif
-
-  char rsa_pem_filename[PATH_MAX];
-  char ecdsa_pem_filename[PATH_MAX];
-  ASSERT_TRUE(createTempFILEpath(rsa_pem_filename));
-  ASSERT_TRUE(createTempFILEpath(ecdsa_pem_filename));
-
-  ScopedFILE rsa_pem(fopen(rsa_pem_filename, "w"));
-  ScopedFILE ecdsa_pem(fopen(ecdsa_pem_filename, "w"));
-  ASSERT_TRUE(rsa_pem);
-  ASSERT_TRUE(ecdsa_pem);
-
-  bssl::UniquePtr<X509> rsa_leaf = GetChainTestCertificate();
-  bssl::UniquePtr<X509> rsa_intermediate = GetChainTestIntermediate();
-  bssl::UniquePtr<X509> ecdsa_leaf = GetECDSATestCertificate();
-  ASSERT_TRUE(PEM_write_X509(rsa_pem.get(), rsa_leaf.get()));
-  ASSERT_TRUE(PEM_write_X509(rsa_pem.get(), rsa_intermediate.get()));
-  ASSERT_TRUE(PEM_write_X509(ecdsa_pem.get(), ecdsa_leaf.get()));
-  rsa_pem.reset();
-  ecdsa_pem.reset();
-
-  bssl::UniquePtr<SSL_CTX> ctx(SSL_CTX_new(TLS_method()));
-  ASSERT_TRUE(ctx);
-  // Load a certificate into |ctx| and verify that |ssl| inherits it.
-  EXPECT_TRUE(SSL_CTX_use_certificate_chain_file(ctx.get(), rsa_pem_filename));
-  bssl::UniquePtr<SSL> ssl(SSL_new(ctx.get()));
-  ASSERT_TRUE(ssl);
-  EXPECT_EQ(X509_cmp(SSL_get_certificate(ssl.get()), rsa_leaf.get()), 0);
-
-  // Load a new cert into |ssl| and verify that it's correctly loaded.
-  EXPECT_TRUE(SSL_use_certificate_chain_file(ssl.get(), ecdsa_pem_filename));
-  EXPECT_EQ(X509_cmp(SSL_get_certificate(ssl.get()), ecdsa_leaf.get()), 0);
-
-  ASSERT_EQ(remove(rsa_pem_filename), 0);
-  ASSERT_EQ(remove(ecdsa_pem_filename), 0);
-}
-
-TEST(SSLTest, IncompatibleTLSVersionState) {
-  // Using the following ASN.1 DER Sequence where 42 is the serialization
-  // format version number of some future version not currently supported:
-  // SEQUENCE {
-  //   SEQUENCE {
-  //     INTEGER { 42 }
-  //   }
-  // }
-  static constexpr size_t INCOMPATIBLE_DER_LEN = 7;
-  static const uint8_t INCOMPATIBLE_DER[INCOMPATIBLE_DER_LEN] = {
-      0x30, 0x05, 0x30, 0x03, 0x02, 0x01, 0x2a};
-
-  bssl::UniquePtr<SSL_CTX> ctx(SSL_CTX_new(TLS_method()));
-  ASSERT_TRUE(ctx);
-
-  ASSERT_FALSE(
-      SSL_from_bytes(INCOMPATIBLE_DER, INCOMPATIBLE_DER_LEN, ctx.get()));
-  ASSERT_EQ(ERR_GET_LIB(ERR_peek_error()), ERR_LIB_SSL);
-  ASSERT_EQ(ERR_GET_REASON(ERR_peek_error()),
-            SSL_R_SERIALIZATION_INVALID_SERDE_VERSION);
-}
-
-// Test that it is possible for the certificate to be configured on a mix of
-// SSL_CTX and SSL. This ensures that we do not inadvertently overshare objects
-// in SSL_new.
-TEST(SSLTest, MixContextAndConnection) {
-  bssl::UniquePtr<SSL_CTX> ctx(SSL_CTX_new(TLS_method()));
-  ASSERT_TRUE(ctx);
-  bssl::UniquePtr<X509> cert = GetTestCertificate();
-  ASSERT_TRUE(cert);
-  bssl::UniquePtr<EVP_PKEY> key = GetTestKey();
-  ASSERT_TRUE(key);
-
-  // Configure the certificate, but not the private key, on the context.
-  ASSERT_TRUE(SSL_CTX_use_certificate(ctx.get(), cert.get()));
-
-  bssl::UniquePtr<SSL> ssl1(SSL_new(ctx.get()));
-  ASSERT_TRUE(ssl1.get());
-  bssl::UniquePtr<SSL> ssl2(SSL_new(ctx.get()));
-  ASSERT_TRUE(ssl2.get());
-
-  // There is no private key configured yet.
-  EXPECT_FALSE(SSL_CTX_get0_privatekey(ctx.get()));
-  EXPECT_FALSE(SSL_get_privatekey(ssl1.get()));
-  EXPECT_FALSE(SSL_get_privatekey(ssl2.get()));
-
-  // Configuring the private key on |ssl1| works.
-  ASSERT_TRUE(SSL_use_PrivateKey(ssl1.get(), key.get()));
-  EXPECT_TRUE(SSL_get_privatekey(ssl1.get()));
-
-  // It does not impact the other connection or the context.
-  EXPECT_FALSE(SSL_CTX_get0_privatekey(ctx.get()));
-  EXPECT_FALSE(SSL_get_privatekey(ssl2.get()));
-}
-
-static size_t test_ecc_privkey_calls = 0;
-
-static enum ssl_private_key_result_t test_ecc_privkey_complete(SSL *ssl,
-                                                           uint8_t *out,
-                                                           size_t *out_len,
-                                                           size_t max_out) {
-  test_ecc_privkey_calls += 1;
-  return ssl_private_key_success;
-}
-
-static enum ssl_private_key_result_t test_ecc_privkey_sign(
-    SSL *ssl, uint8_t *out, size_t *out_len, size_t max_out,
-    uint16_t signature_algorithm, const uint8_t *in, size_t in_len) {
-  bssl::UniquePtr<EVP_PKEY> pkey(GetECDSATestKey());
-
-  if (EVP_PKEY_id(pkey.get()) !=
-      SSL_get_signature_algorithm_key_type(signature_algorithm)) {
-      return ssl_private_key_failure;
-  }
-
-  const EVP_MD *md = SSL_get_signature_algorithm_digest(signature_algorithm);
-  bssl::ScopedEVP_MD_CTX ctx;
-  EVP_PKEY_CTX *pctx = nullptr;
-  if (!EVP_DigestSignInit(ctx.get(), &pctx, md, nullptr,
-                          pkey.get())) {
-    return ssl_private_key_failure;
-  }
-
-  size_t len = 0;
-  if (!EVP_DigestSign(ctx.get(), nullptr, &len, in, in_len) || len > max_out) {
-    return ssl_private_key_failure;
-  }
-
-  *out_len = max_out;
-
-  if (!EVP_DigestSign(ctx.get(), out, out_len, in, in_len)) {
-    return ssl_private_key_failure;
-  }
-
-  return test_ecc_privkey_complete(ssl, out, out_len, max_out);
-}
-
-static enum ssl_private_key_result_t test_ecc_privkey_decrypt(
-    SSL *ssl, uint8_t *out, size_t *out_len, size_t max_out, const uint8_t *in,
-    size_t in_len) {
-  return ssl_private_key_failure;
-}
-
-static const SSL_PRIVATE_KEY_METHOD test_ecc_private_key_method = {
-    test_ecc_privkey_sign,
-    test_ecc_privkey_decrypt,
-    test_ecc_privkey_complete,
-};
-
-TEST(SSLTest, SSLPrivateKeyMethod) {
-  {
-    bssl::UniquePtr<SSL_CTX> client_ctx(SSL_CTX_new(TLS_method()));
-    bssl::UniquePtr<SSL_CTX> server_ctx(SSL_CTX_new(TLS_method()));
-
-    bssl::UniquePtr<X509> ecdsa_cert(GetECDSATestCertificate());
-    bssl::UniquePtr<CRYPTO_BUFFER> ecdsa_leaf =
-        x509_to_buffer(ecdsa_cert.get());
-    std::vector<CRYPTO_BUFFER *> chain = {
-        ecdsa_leaf.get(),
-    };
-
-    // Index should be have been set to default, 0, but no key loaded
-    EXPECT_EQ(server_ctx->cert->cert_private_key_idx, SSL_PKEY_RSA);
-    EXPECT_EQ(
-        server_ctx->cert->cert_private_keys[SSL_PKEY_RSA].privatekey.get(),
-        nullptr);
-    EXPECT_EQ(server_ctx->cert->key_method, nullptr);
-
-
-    // Load a certificate chain containg the leaf but set private key method
-    ASSERT_TRUE(SSL_CTX_set_chain_and_key(server_ctx.get(), &chain[0],
-                                          chain.size(), nullptr,
-                                          &test_ecc_private_key_method));
-
-    // Should be initiall zero
-    ASSERT_EQ(test_ecc_privkey_calls, (size_t)0);
-
-    // Index must be ECC key now, but key_method must be set.
-    ASSERT_EQ(server_ctx->cert->cert_private_key_idx, SSL_PKEY_ECC);
-    ASSERT_EQ(server_ctx->cert->key_method, &test_ecc_private_key_method);
-
-    bssl::UniquePtr<SSL> client, server;
-    ASSERT_TRUE(ConnectClientAndServer(&client, &server, client_ctx.get(),
-                                       server_ctx.get(), ClientConfig(),
-                                       false));
-
-    ASSERT_TRUE(CompleteHandshakes(client.get(), server.get()));
-
-    ASSERT_EQ(test_ecc_privkey_calls, (size_t)1);
-
-    // Check the internal slot index to verify that the correct slot was used
-    // during the handshake.
-    ASSERT_EQ(server->config->cert->cert_private_key_idx, SSL_PKEY_ECC);
-    ASSERT_EQ(server->config->cert->key_method, &test_ecc_private_key_method);
-  }
-
-  {
-    size_t current_invoke_count = test_ecc_privkey_calls;
-
-    bssl::UniquePtr<SSL_CTX> client_ctx(SSL_CTX_new(TLS_method()));
-    bssl::UniquePtr<SSL_CTX> server_ctx(SSL_CTX_new(TLS_method()));
-
-    // Index should be have been set to default, 0, but no key loaded
-    EXPECT_EQ(server_ctx->cert->cert_private_key_idx, SSL_PKEY_RSA);
-    EXPECT_EQ(
-        server_ctx->cert->cert_private_keys[SSL_PKEY_RSA].privatekey.get(),
-        nullptr);
-    EXPECT_EQ(server_ctx->cert->key_method, nullptr);
-
-    bssl::UniquePtr<X509> ed_cert(GetED25519TestCertificate());
-    bssl::UniquePtr<EVP_PKEY> ed_key(GetED25519TestKey());
-    bssl::UniquePtr<CRYPTO_BUFFER> ed_leaf = x509_to_buffer(ed_cert.get());
-    std::vector<CRYPTO_BUFFER *> ed_chain = {
-        ed_leaf.get(),
-    };
-
-    // Load a certificate chain containg the leaf but set private key method
-    ASSERT_TRUE(SSL_CTX_set_chain_and_key(server_ctx.get(), &ed_chain[0],
-                                          ed_chain.size(), ed_key.get(),
-                                          nullptr));
-
-    // Index must be ECC key now, but key_method must not be set.
-    ASSERT_EQ(server_ctx->cert->cert_private_key_idx, SSL_PKEY_ED25519);
-    ASSERT_EQ(server_ctx->cert->key_method, nullptr);
-
-    std::vector<uint16_t> sigalgs = {SSL_SIGN_ED25519};
-
-    ASSERT_TRUE(SSL_CTX_set_signing_algorithm_prefs(
-        client_ctx.get(), sigalgs.data(), sigalgs.size()));
-    ASSERT_TRUE(SSL_CTX_set_verify_algorithm_prefs(
-        client_ctx.get(), sigalgs.data(), sigalgs.size()));
-
-    bssl::UniquePtr<SSL> client, server;
-    ASSERT_TRUE(ConnectClientAndServer(&client, &server, client_ctx.get(),
-                                       server_ctx.get(), ClientConfig(),
-                                       false));
-
-    ASSERT_TRUE(CompleteHandshakes(client.get(), server.get()));
-
-    // This should still be the same, as we didn't use the private key method
-    // functionality, so it shouldn't have incremented.
-    ASSERT_EQ(test_ecc_privkey_calls, current_invoke_count);
-
-    // Check the internal slot index to verify that the correct slot was used
-    // during the handshake and that key_method was not set.
-    ASSERT_EQ(server->config->cert->cert_private_key_idx, SSL_PKEY_ED25519);
-    ASSERT_EQ(server->config->cert->key_method, nullptr);
-  }
 }
 
 }  // namespace
